### 2021-04-03 #37 @arianesasso
Devicely: A Python package for reading, timeshifting and writing sensor data
Submitting Author: Ariane Sasso (@arianesasso)
Package Name: Devicely
One-Line Description of Package: A Python package for reading, timeshifting and writing sensor data
Repository Link: https://github.com/hpi-dhc/devicely
Version submitted: 0.2.5
Version accepted: TBD

---

Wearable devices can track a multitude of parameters such as heart rate, body temperature, blood oxygen saturation, acceleration, blood glucose and much more [Kamisalic2018]. Moreover, they are becoming increasingly popular with a steeping increase in market presence in 2020 alone [IDC2020]. Applications for wearable devices varies from tracking cardiovascular risks [Bayoumy2021] to identifying COVID-19 onset [Mishra2020]. Therefore, there is a great need for scientists to easily go through data acquired from different wearables. In order to solve this problem and empower scientists working with biosignals, we developed the devicely package. It represents the data in a science-friendly format and lets scientists focus on what they want: the analysis of biosignals.

\* Please fill out a pre-submission inquiry before submitting a data visualization package. For more info, see [notes on categories][NotesOnCategories] of our guidebook.

- Explain how the and why the package falls under these categories (briefly, 1-2 sentences):
We developed this package for scientists to easily go through data acquired from different wearables.

Scientists working with wearable devices.

Not to our knowledge.

- [x] has an [OSI approved license][OsiApprovedLicense].

- [x] The package is deposited in a long-term repository with the DOI: 10.5281/zenodo.4661545

- [x] I agree to abide by [pyOpenSci's Code of Conduct][PyOpenSciCodeOfConduct] during the review process and in maintaining my package should it be accepted.

[Editor and review templates can be found here][Templates]

[NotesOnCategories]: https://www.pyopensci.org/contributing-guide/open-source-software-peer-review/aims-and-scope.html?highlight=data#notes-on-categories

[Templates]: https://www.pyopensci.org/contributing-guide/appendices/templates.html
