+++ README.md
@@ -1,4 +1,5 @@
 [![Actions Status: test](https://github.com/hpi-dhc/devicely/workflows/test/badge.svg)](https://github.com/hpi-dhc/devicely/actions/workflows/test.yml)
+[![DOI](https://zenodo.org/badge/279395106.svg)](https://zenodo.org/badge/latestdoi/{279395106})
 
 ![Devicely Logo](/imgs/logo/devicely-logo.png)
 
