### 2021-04-30 #37 @lwasser
this package is definitely in scope for pyopensci review!.
