### 2021-05-21 #37 @xmnlab
@arianesasso @lwasser, adding initial editor checks here.

@arianesasso we are lining up reviewers for your package. When we receive the confirmation I will update the text here with this information.

## Editor checks:

- [x] **License:** The package has an OSI accepted license

---

## Editor comments

---

Reviewers:
Due date:
