### 2021-05-24 #37 @xmnlab
Hi @willingc! @lwasser told me that she has talked to you about reviewing a package for pyopensci! that sounds great!

PyOpenSci contributing guide is here https://www.pyopensci.org/contributing-guide/intro.html

This is the guide specifically for reviewers: https://www.pyopensci.org/contributing-guide/open-source-software-submissions/reviewer-guide.html

This is my first time acting as editor, so I still in the learning curve process :) I have reviewed some packages here so feel free to ping me anytime you need any information or help.

I am waiting for the confirmation of one more reviewer and after that, I will update the schedules for this review.

thank you so much for your support :)
