### 2021-05-25 #37 @xmnlab
Hi @agricolab!

@NickleDave told me you are available to review this package, thank you so much!

PyOpenSci contributing guide is here https://www.pyopensci.org/contributing-guide/intro.html

And this is the guide specifically for reviewers: https://www.pyopensci.org/contributing-guide/open-source-software-submissions/reviewer-guide.html

As I mentioned before, this is my first time acting as an editor, so I still in the learning curve process :) I have reviewed some packages here so feel free to ping me anytime you need any information or help.

I am updating today some information in the first comment in this issue, including the due date for the reviews.

thank you so much for your support!
