### 2021-05-25 #37 @xmnlab
@arianesasso @agricolab @willingc @lwasser

I updated the information here
https://github.com/pyOpenSci/software-review/issues/37#issuecomment-846220959

Let me know if I can help in any way! thank you all!
