### 2021-05-26 #37 @willingc
@xmnlab I started last night and made some notes in a [gist](https://gist.github.com/willingc/988c6af3ff0ba8c14769a1127f4b1db4)
