### 2021-06-01 hpi-dhc/devicely#19 @agricolab
Many tests fail when running locally with outdated dependencies
I have just cloned the repo and installed the package. When i run pytest similar to how it is run in the github action, most tests fail.
```
pytest -rA --ignore=tests/test_faros.py

...

============================================ short test summary info =============================================
PASSED tests/test_empatica.py::EmpaticaTestCase::test_write
PASSED tests/test_spacelabs.py::SpacelabsTestCase::test_deidentify
PASSED tests/test_spacelabs.py::SpacelabsTestCase::test_drop_EB
PASSED tests/test_spacelabs.py::SpacelabsTestCase::test_random_timeshift
PASSED tests/test_spacelabs.py::SpacelabsTestCase::test_read
PASSED tests/test_spacelabs.py::SpacelabsTestCase::test_write
PASSED tests/test_tags.py::TagsTestCase::test_random_timeshift
PASSED tests/test_tags.py::TagsTestCase::test_read
PASSED tests/test_tags.py::TagsTestCase::test_write
FAILED tests/test_empatica.py::EmpaticaTestCase::test_joined_dataframe - TypeError: assert_frame_equal() got an...
FAILED tests/test_empatica.py::EmpaticaTestCase::test_read - TypeError: assert_frame_equal() got an unexpected ...
FAILED tests/test_everion.py::EverionTestCase::test_basic_read - KeyError: "[('count', '<lambda>')] not in index"
FAILED tests/test_everion.py::EverionTestCase::test_basic_write - KeyError: "[('count', '<lambda>')] not in index"
FAILED tests/test_everion.py::EverionTestCase::test_random_timeshift - KeyError: "[('count', '<lambda>')] not i...
FAILED tests/test_everion.py::EverionTestCase::test_read_with_all_join_files_missing - KeyError: "[('count', '<...
FAILED tests/test_everion.py::EverionTestCase::test_read_with_invalid_tags - KeyError: "[('count', '<lambda>')]...
FAILED tests/test_everion.py::EverionTestCase::test_read_with_missing_files - KeyError: "[('count', '<lambda>')...
FAILED tests/test_everion.py::EverionTestCase::test_read_with_non_default_tags - KeyError: "[('count', '<lambda...
FAILED tests/test_everion.py::EverionTestCase::test_timeshift_by_timedelta - KeyError: "[('count', '<lambda>')]...
FAILED tests/test_everion.py::EverionTestCase::test_timeshift_to_timestamp - KeyError: "[('count', '<lambda>')]...
FAILED tests/test_shimmer.py::ShimmerTestCase::test_basic_read - TypeError: an integer is required (got type str)
FAILED tests/test_shimmer.py::ShimmerTestCase::test_random_timeshift - TypeError: an integer is required (got t...
FAILED tests/test_shimmer.py::ShimmerTestCase::test_timeshift_by_timedelta - TypeError: an integer is required ...
FAILED tests/test_shimmer.py::ShimmerTestCase::test_timeshift_to_timestamp - TypeError: an integer is required ...
FAILED tests/test_shimmer.py::ShimmerTestCase::test_write - TypeError: an integer is required (got type str)
FAILED tests/test_spacelabs.py::SpacelabsTestCase::test_set_window_column - KeyError: "None of [Index(['window_...
FAILED tests/test_spacelabs.py::SpacelabsTestCase::test_set_window_index - AssertionError: Timestamp('1999-01-0...
```
when i created a fresh environment and installed all dependencies fresh, all tests passed. The versions of the dependencies under which the local tests passed where (according to pip list):
<pre>Package            Version
------------------ --------
devicely           0.2.5
matplotlib         3.4.2
numpy              1.20.3
pandas             1.2.4
pyEDFlib           0.1.22
pytest             6.2.4
scipy              1.6.3
xmltodict          0.12.0

</pre>
A solution might be pinning the required versions of your dependencies.
