### 2021-06-01 #37 @arianesasso
> @xmnlab is this a co-submission to JOSS?

@agricolab that was our goal :), the paper is in inst/paper.md
