+++ README.md
@@ -1,5 +1,6 @@
 [![Actions Status: test](https://github.com/hpi-dhc/devicely/workflows/test/badge.svg)](https://github.com/hpi-dhc/devicely/actions/workflows/test.yml)
 [![DOI](https://zenodo.org/badge/279395106.svg)](https://zenodo.org/badge/latestdoi/{279395106})
+[![Project Status: Active – The project has reached a stable, usable state and is being actively developed.](https://www.repostatus.org/badges/latest/active.svg)](https://www.repostatus.org/#active)
 
 ![Devicely Logo](/imgs/logo/devicely-logo.png)
 
