### 2021-06-12 #37 @arianesasso
Dear @agricolab, thanks again for the great review! We added a repostatus.org badge and will add another one with test coverage (possibly https://github.com/nedbat/coveragepy; other suggestions are welcomed as well). We will fix hpi-dhc/devicely#19. Also, thanks for pointing out the issue with FAROS in CI/tests (we are investigating). More, we will add HISTORY/CHANGELOG.md which is indeed missing in the root directory; and we are currently writing a new section in the paper.md with the comparison to other packages (great point). We will notify everyone when the changes are ready :).
