+++ .github/workflows/test.yml
@@ -2,37 +2,47 @@ name: test
 
 on:
   push:
-    branches:
-    - master
-  pull_request:
-    branches:
-    - master
+  #   branches:
+  #   - master
+  # pull_request:
+  #   branches:
+  #   - master
 
 jobs:
   build:
-
     runs-on: ubuntu-latest
     strategy:
       matrix:
-        python-version: [3.6, 3.7, 3.8, 3.9]
-
+        python-version: [3.9]
     steps:
     - uses: actions/checkout@v2
+    
     - name: Set up Python ${{ matrix.python-version }}
       uses: actions/setup-python@v2
       with:
         python-version: ${{ matrix.python-version }}
+    
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
-        pip install flake8 pytest
+        pip install pytest coverage
         pip install .
-    - name: Lint with flake8
-      run: |
-        # stop the build if there are Python syntax errors or undefined names
-        flake8 . --count --select=E9,F63,F7,F82 --show-source --statistics
-        # exit-zero treats all errors as warnings. The GitHub editor is 127 chars wide
-        flake8 . --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
+    
     - name: Test with pytest
       run: |
-        python -m pytest --ignore=tests/test_faros.py
+        coverage run --source test -m pytest --ignore=tests/test_faros.py
+
+    - if: ${{ matrix.python-version == '3.9' }}
+      name: Create the Coverage Badge
+      run: |
+        echo "COVERAGE=$(python -c "import xml.etree.ElementTree as ET; print(int(float(ET.parse('coverage.xml').getroot().attrib['line-rate']) * 100))")" >> $GITHUB_ENV
+        echo "COLOR=$(echo $COVERAGE | python -c "import sys; from bisect import bisect; i=bisect([0,60,70,80,80,100], int(sys.stdin.read()))-1; print(['red', 'orange', 'yellow', 'yellowgreen', 'green', 'brightgreen'][i])")" >> $GITHUB_ENV
+
+      uses: schneegans/dynamic-badges-action@v1.0.0
+      with:
+        auth: ${{ secrets.GIST_SECRET }}
+        gistID: 270a0114dfad9251945a146dd6d29fa6
+        filename: devicely_coverage__${{ env.BRANCH }}.json
+        label: Test Coverage
+        message: ${{ env.COVERAGE }}
+        color: ${{ env.COLOR }}
