+++ .github/workflows/test.yml
@@ -33,11 +33,13 @@ jobs:
         coverage run --source test -m pytest --ignore=tests/test_faros.py
 
     - if: ${{ matrix.python-version == '3.9' }}
-      name: Create the Coverage Badge
+      name: Calculate coverage and color
       run: |
         echo "COVERAGE=$(python -c "import xml.etree.ElementTree as ET; print(int(float(ET.parse('coverage.xml').getroot().attrib['line-rate']) * 100))")" >> $GITHUB_ENV
         echo "COLOR=$(echo $COVERAGE | python -c "import sys; from bisect import bisect; i=bisect([0,60,70,80,80,100], int(sys.stdin.read()))-1; print(['red', 'orange', 'yellow', 'yellowgreen', 'green', 'brightgreen'][i])")" >> $GITHUB_ENV
 
+    - if: ${{ matrix.python-version == '3.9' }}
+      name: Create the coverage gist
       uses: schneegans/dynamic-badges-action@v1.0.0
       with:
         auth: ${{ secrets.GIST_SECRET }}
