### 2021-06-26 d91856a
change coverage command in workflow
