+++ .github/workflows/test.yml
@@ -32,6 +32,11 @@ jobs:
       run: |
          coverage run --source=tests -m pytest --ignore=tests/test_faros.py
 
+    - name: Test if coverage.xml exists
+      run: |
+        ls -lisah
+        find.
+
     - if: ${{ matrix.python-version == '3.9' }}
       name: Calculate coverage and color
       run: |
