### 2021-06-26 d7ddd67
add line to workflow that creates coverage.xml
