+++ .github/workflows/test.yml
@@ -30,9 +30,10 @@ jobs:
     
     - name: Test with pytest
       run: |
-         coverage run --source=tests -m pytest --ignore=tests/test_faros.py
+         coverage run --source=tests -m pytest tests/test_spacelabs.py
          coverage xml
-         head -n 30 coverage.xml
+         python -c "import xml.etree.ElementTree as ET; print(int(float(ET.parse('coverage.xml').getroot().attrib['line-rate']) * 100))"
+         python -c "import sys; from bisect import bisect; i=bisect([0,60,70,80,80,100], int(sys.stdin.read()))-1; print(['red', 'orange', 'yellow', 'yellowgreen', 'green', 'brightgreen'][i])"
 
     - if: ${{ matrix.python-version == '3.9' }}
       name: Calculate coverage and color
