+++ .github/workflows/test.yml
@@ -32,8 +32,9 @@ jobs:
       run: |
          coverage run --source=tests -m pytest tests/test_spacelabs.py
          coverage xml
-         python -c "import xml.etree.ElementTree as ET; print(int(float(ET.parse('coverage.xml').getroot().attrib['line-rate']) * 100))"
-         python -c "import sys; from bisect import bisect; i=bisect([0,60,70,80,80,100], int(sys.stdin.read()))-1; print(['red', 'orange', 'yellow', 'yellowgreen', 'green', 'brightgreen'][i])"
+         head -n 2 coverage.xml
+         python -c "import xml.etree.ElementTree as ET; print(ET.parse('coverage.xml').getroot().attrib['line-rate'])"
+  #python -c "import sys; from bisect import bisect; i=bisect([0,60,70,80,80,100], int(sys.stdin.read()))-1; print(['red', 'orange', 'yellow', 'yellowgreen', 'green', 'brightgreen'][i])"
 
     - if: ${{ matrix.python-version == '3.9' }}
       name: Calculate coverage and color
