+++ .github/workflows/test.yml
@@ -39,12 +39,6 @@ jobs:
          echo $COLOR
 
     - if: ${{ matrix.python-version == '3.9' }}
-      name: Calculate coverage and color
-      run: |
-        echo "COVERAGE=$(python -c "import xml.etree.ElementTree as ET; print(int(float(ET.parse('coverage.xml').getroot().attrib['line-rate']) * 100))")" >> $GITHUB_ENV
-        echo "COLOR=$(echo $COVERAGE | python -c "import sys; from bisect import bisect; i=bisect([0,60,70,80,80,100], int(sys.stdin.read()))-1; print(['red', 'orange', 'yellow', 'yellowgreen', 'green', 'brightgreen'][i])")" >> $GITHUB_ENV
-
-    - if: ${{ matrix.python-version == '3.9' }}
       name: Create the coverage gist
       uses: schneegans/dynamic-badges-action@v1.0.0
       with:
@@ -52,5 +46,5 @@ jobs:
         gistID: 270a0114dfad9251945a146dd6d29fa6
         filename: devicely_coverage__${{ env.BRANCH }}.json
         label: Test Coverage
-        message: ${{ env.COVERAGE }}
-        color: ${{ env.COLOR }}
+        message: $COVERAGE
+        color: $COLOR
