+++ .github/workflows/test.yml
@@ -44,7 +44,7 @@ jobs:
       with:
         auth: ${{ secrets.GIST_SECRET }}
         gistID: 270a0114dfad9251945a146dd6d29fa6
-        filename: devicely_coverage__${{ env.BRANCH }}.json
+        filename: devicely__${{ env.BRANCH }}.json
         label: Test Coverage
-        message: $COVERAGE
-        color: $COLOR
+        message: ${{ COVERAGE }}
+        color: ${{ COLOR }}
