+++ .github/workflows/test.yml
@@ -37,6 +37,8 @@ jobs:
          echo $COVERAGE
          COLOR=$(echo $COVERAGE | python -c "import sys; from bisect import bisect; i=bisect([0,60,70,80,80,100], int(sys.stdin.read()))-1; print(['red', 'orange', 'yellow', 'yellowgreen', 'green', 'brightgreen'][i])")
          echo $COLOR
+         echo "COVERAGE=$COVERAGE" >> $GITHUB_ENV
+         echo "COLOR=$COLOR" >> $GITHUB_ENV
 
     - if: ${{ matrix.python-version == '3.9' }}
       name: Create the coverage gist
@@ -46,5 +48,5 @@ jobs:
         gistID: 270a0114dfad9251945a146dd6d29fa6
         filename: devicely__${{ env.BRANCH }}.json
         label: Test Coverage
-        message: ${{ COVERAGE }}
-        color: ${{ COLOR }}
+        message: ${{ env.COVERAGE }}
+        color: ${{ env.COLOR }}
