+++ .github/workflows/test.yml
@@ -1,19 +1,19 @@
-name: test
+name: Testing and coverage
 
 on:
   push:
-  #   branches:
-  #   - master
-  # pull_request:
-  #   branches:
-  #   - master
+    branches:
+    - master
+  pull_request:
+    branches:
+    - master
 
 jobs:
   build:
     runs-on: ubuntu-latest
     strategy:
       matrix:
-        python-version: [3.9]
+        python-version: [3.6, 3.7, 3.8, 3.9]
     steps:
     - uses: actions/checkout@v2
     
@@ -30,13 +30,15 @@ jobs:
     
     - name: Test with pytest
       run: |
+         python -m pytest tests/test_spacelabs.py
+
+    - if: ${{ matrix.python-version == '3.9' }}
+      name: Calculate coverage
+      run: |
          coverage run --source=tests -m pytest tests/test_spacelabs.py
          coverage xml
-         head -n 2 coverage.xml
          COVERAGE=$(python -c "import xml.etree.ElementTree as ET; print(int(float(ET.parse('coverage.xml').getroot().attrib['line-rate']) * 100))")
-         echo $COVERAGE
          COLOR=$(echo $COVERAGE | python -c "import sys; from bisect import bisect; i=bisect([0,60,70,80,80,100], int(sys.stdin.read()))-1; print(['red', 'orange', 'yellow', 'yellowgreen', 'green', 'brightgreen'][i])")
-         echo $COLOR
          echo "COVERAGE=$COVERAGE" >> $GITHUB_ENV
          echo "COLOR=$COLOR" >> $GITHUB_ENV
 
