### 2021-06-26 d9c3848
edit workflow to test all modules
