+++ .github/workflows/test.yml
@@ -30,12 +30,12 @@ jobs:
     
     - name: Test with pytest
       run: |
-         python -m pytest tests/test_spacelabs.py
+         python -m pytest
 
     - if: ${{ matrix.python-version == '3.9' }}
       name: Calculate coverage
       run: |
-         coverage run --source=tests -m pytest tests/test_spacelabs.py
+         coverage run --source=tests -m pytest
          coverage xml
          COVERAGE=$(python -c "import xml.etree.ElementTree as ET; print(int(float(ET.parse('coverage.xml').getroot().attrib['line-rate']) * 100))")
          COLOR=$(echo $COVERAGE | python -c "import sys; from bisect import bisect; i=bisect([0,60,70,80,80,100], int(sys.stdin.read()))-1; print(['red', 'orange', 'yellow', 'yellowgreen', 'green', 'brightgreen'][i])")
