### 2021-06-26 5b61501
change bounds for coverage badge color selection in workflow
