### 2021-06-28 affc2cf
rewrite code in devicely/faros.py
