+++ devicely/faros.py
@@ -1,33 +1,26 @@
 """
 Module to process Faros 180 data
 """
-import json
-import os
+from itertools import chain
+from posixpath import join
 import random
 
 import numpy as np
 import pandas as pd
-import pyedflib
+import pyedflib as edf
 
 
 class FarosReader:
     """
     Parses, timeshifts and writes data
     generated by Bittium Faros devices.
-    This reader reads data from .edf files and writes it back a .csv file.
-    It can read back data from a .csv file if it was written using this reader class.
 
     Attributes
     ----------
     data : DataFrame
         DataFrame with the values that were read from the .edf or .csv file.
-
-    start_time : pandas.Timestamp
-        global start time of all measurements
-
-    sample_freqs : dict
-        sampling frequencies of all columns
     """
+
     def __init__(self, path):
         """
         Reads an .edf file generated by the device or a .csv file generated by this reader.
@@ -37,69 +30,89 @@ class FarosReader:
         path : str
             Path of the .edf or .csv file.
         """
-        self.start_time = 0
-        file_extension = os.path.splitext(path)[-1].lower()
-        if file_extension == '.edf':
-            self._read_from_edf(path)
-        elif file_extension == '.csv':
-            self._read_from_csv(path)
-        else:
-            raise ValueError(f"Wrong file extension. Expected one of [.edf, .csv]. Got {file_extension}.")
+        reader = edf.EdfReader(path)
+        self.file_header = reader.getHeader()
+        self.signal_headers = reader.getSignalHeaders()
+        self.sample_freqs = {header['label']: header['sample_rate'] 
+                             for header in self.signal_headers}
+        self.start_time = pd.Timestamp(reader.getHeader()['startdate'])
+
+        self._n_samples = reader.getNSamples()
+        self._n_datarecords = reader.datarecords_in_file
+
+        self.ECG = pd.Series(
+            reader.readSignal(0),
+            name='ECG', 
+            index=pd.date_range(start=self.start_time,
+                                periods=self._n_samples[0],
+                                freq=f"{1/self.sample_freqs['ECG']}S")
+        )
+        self.ACC = pd.DataFrame({'X': reader.readSignal(1),
+                                 'Y': reader.readSignal(2),
+                                 'Z': reader.readSignal(3)},
+                                 index=pd.date_range(start=self.start_time,
+                                                     periods=self._n_samples[1],
+                                                     freq=f"{1/self.sample_freqs['Accelerometer_X']}S")
+        )
+        self.ACC['mag'] = np.linalg.norm(self.ACC, axis=1)
+        self.Marker = pd.Series(
+            reader.readSignal(4),
+            name='Marker',
+            index=pd.date_range(start=self.start_time,
+                                periods=self._n_samples[4],
+                                freq=f"{1/self.sample_freqs['Accelerometer_X']}S")
+        )
+        self.HRV = pd.Series(
+            reader.readSignal(5),
+            name='HRV',
+            index=pd.date_range(start=self.start_time,
+                                periods=self._n_samples[5],
+                                freq=f"{1/self.sample_freqs['HRV']}S")
+        )
+
+        reader.close()
+
+        self.data = self._get_joined_df()
+
+        a = 0
+
+    def _get_joined_df(self):
+        # get index for joined dataframe
+        joined_idx = pd.concat(map(pd.Series, [self.ECG.index, self.ACC.index, self.Marker.index, self.HRV.index]))
[…]
