### 2021-06-28 e445106
made faros->write to edf work
