+++ .github/workflows/test.yml
@@ -3,10 +3,10 @@ name: Testing and coverage
 on:
   push:
     branches:
-    - master
+    - main
   pull_request:
     branches:
-    - master
+    - main
 
 jobs:
   build:
@@ -30,12 +30,12 @@ jobs:
     
     - name: Test with pytest
       run: |
-         python -m pytest
+         python -m pytest --ignore=tests/test_faros.py
 
     - if: ${{ matrix.python-version == '3.9' }}
       name: Calculate coverage
       run: |
-         coverage run --source=tests -m pytest
+         coverage run --source=tests -m pytest --ignore=tests/test_faros.py
          coverage xml
          COVERAGE=$(python -c "import xml.etree.ElementTree as ET; print(int(float(ET.parse('coverage.xml').getroot().attrib['line-rate']) * 100))")
          COLOR=$(echo $COVERAGE | python -c "import sys; from bisect import bisect; i=bisect([0,60,70,80,95,100], int(sys.stdin.read()))-1; print(['red', 'orange', 'yellow', 'yellowgreen', 'green', 'brightgreen'][i])")
@@ -48,7 +48,7 @@ jobs:
       with:
         auth: ${{ secrets.GIST_SECRET }}
         gistID: 270a0114dfad9251945a146dd6d29fa6
-        filename: devicely_coverage_master.json
+        filename: devicely_coverage_main.json
         label: Test Coverage
         message: ${{ env.COVERAGE }}%
         color: ${{ env.COLOR }}
