### 2021-06-28 e453c7b
parse spacelabs metadata using ETree instead of xmltodict
