### 2021-06-29 #37 @xmnlab
hi @arianesasso ! as the last update here was 2 weeks ago I would like to check about the progress. let us know if you have any block or if you want to discuss any point related to the review :)
