### 2021-07-01 #37 @arianesasso
Dear @xmnlab, thanks for checking! We are implementing the requests of the reviewers, e. g. we have the new badges and other small changes already in there.

Currently, when checking the problem with the tests: `Faros is apparently ignored during CI/ automated tests? When i run it locally, it works, though. Please explain the reason for ignoring it, though, so i can consider whether the item "automated tests" is fullfilled.`

We found out why it was being ignored and decided this required some changes in the code to make it more general (and this is taking some time). I think we will be able to have everything ready in 2 weeks. I hope that is ok.
