### 2021-07-18 589c098
implemented reading from and writing to both edf and directory in faros
