+++ devicely/faros.py
@@ -2,6 +2,8 @@
 Module to process Faros 180 data
 """
 from itertools import chain
+import json
+import os
 from posixpath import join
 import random
 
@@ -23,64 +25,111 @@ class FarosReader:
 
     def __init__(self, path):
         """
-        Reads an .edf file generated by the device or a .csv file generated by this reader.
+        Reads an Faros-generated EDF file or a directory created by this class.
 
         Parameters
         ----------
         path : str
-            Path of the .edf or .csv file.
+            Can either be Faros-generated EDF file or a directory created by the FarosReader.write() method.
         """
+
+        self.start_time = None
+        self.sample_freqs = None
+        self.units = None
+        self.edf_metadata = None
+        self.ECG = None
+        self.ACC = None
+        self.Marker = None
+        self.HRV = None
+
+        if os.path.isfile(path):
+            self._read_from_edf_file(path)
+        if os.path.isdir(path):
+            self._read_from_directory(path)
+
+        self.data = self._get_joined_df()
+
+    def _read_from_edf_file(self, path):
         reader = edf.EdfReader(path)
-        self.file_header = reader.getHeader()
-        self.signal_headers = reader.getSignalHeaders()
-        self.sample_freqs = {header['label']: header['sample_rate'] 
-                             for header in self.signal_headers}
-        self.start_time = pd.Timestamp(reader.getHeader()['startdate'])
+        self.start_time = pd.Timestamp(reader.getStartdatetime())
+        self.sample_freqs = {
+            'ECG': reader.getSampleFrequency(0),
+            'ACC': reader.getSampleFrequency(1),
+            'Marker': reader.getSampleFrequency(4),
+            'HRV': reader.getSampleFrequency(5)
+        }
+        self.units = {
+            'ECG': reader.getSignalHeader(0)['dimension'],
+            'ACC': reader.getSignalHeader(1)['dimension'],
+            'HRV': reader.getSignalHeader(5)['dimension'],
+        }
+        self.edf_metadata = reader.getSignalHeaders()
 
         self._n_samples = reader.getNSamples()
         self._n_datarecords = reader.datarecords_in_file
 
-        self.ECG = pd.Series(
-            reader.readSignal(0),
-            name='ECG', 
-            index=pd.date_range(start=self.start_time,
-                                periods=self._n_samples[0],
-                                freq=f"{1/self.sample_freqs['ECG']}S")
+        ecg = reader.readSignal(0)
+        self.ECG = pd.Series(ecg, name='ECG',
+                             index=pd.date_range(start=self.start_time,
+                                                 periods=len(ecg),
+                                                 freq=f"{1/self.sample_freqs['ECG']}S")
         )
-        self.ACC = pd.DataFrame({'X': reader.readSignal(1),
-                                 'Y': reader.readSignal(2),
-                                 'Z': reader.readSignal(3)},
-                                 index=pd.date_range(start=self.start_time,
-                                                     periods=self._n_samples[1],
-                                                     freq=f"{1/self.sample_freqs['Accelerometer_X']}S")
+
+        acc = np.array([reader.readSignal(i) for i in range(1, 4)]).T
+        self.ACC = pd.DataFrame(acc, columns=['X', 'Y', 'Z'],
+                                     index=pd.date_range(start=self.start_time,
+                                                         periods=len(acc),
[…]
