+++ tests/test_faros.py
@@ -1,95 +1,260 @@
 import os
+import shutil
 import unittest
 
 import numpy as np
 import pandas as pd
 
-import sys
-sys.path.insert(0, '/home/jost/DHC/devicely')
+
 import devicely
 
 class FarosTestCase(unittest.TestCase):
-    READ_PATH = 'tests/Faros_test_data/testfilename.EdF'
-    WRITE_PATH = 'tests/Faros_test_data/writtentestfile.csv'
-
+    EDF_READ_PATH = 'tests/Faros_test_data/testfile_read.EDF'
+    EDF_WRITE_PATH = 'tests/Faros_test_data/testfile_write.EDF'
+    DIR_READ_PATH = 'tests/Faros_test_data/testdir_read'
+    DIR_WRITE_PATH = 'tests/Faros_test_data/testdir_write'
 
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
-        self.reader = devicely.FarosReader(self.READ_PATH)
-
-    def setUp(self):
+        self.reader_from_edf = devicely.FarosReader(self.EDF_READ_PATH)
+        self.reader_from_dir = devicely.FarosReader(self.DIR_READ_PATH)
+        
+        self.expected_start_time = pd.Timestamp('2018-10-12 16:54:12')
         self.expected_sample_freqs = {
-            'ECG': pd.DateOffset(seconds=1/1000),
-            'Accelerometer_X': pd.DateOffset(seconds=1/100),
-            'Accelerometer_Y': pd.DateOffset(seconds=1/100),
-            'Accelerometer_Z': pd.DateOffset(seconds=1/100),
-            'Marker': pd.DateOffset(seconds=1/1),
-            'HRV': pd.DateOffset(seconds=1/5),
-            'acc_mag': pd.DateOffset(seconds=1/100)
+            'ECG': 500.0,
+            'ACC': 25.0,
+            'Marker': 1.0,
+            'HRV': 5.0
         }
-        self.expected_signal_heads = {
-            'ECG': np.array([ 26.,  -6., -31., -39., -17.]),
-            'Accelerometer_X': np.array([164., 152., 152., 117., -47.]),
-            'Accelerometer_Y': np.array([ 23.,  23., -24.,  11., 246.]),
-            'Accelerometer_Z': np.array([-1172., -1172., -1079.,  -985., -1125.]),
-            'Marker': np.array([0., 0., 0., 0., 0.]),
-            'HRV': np.array([0., 0., 0., 0., 0.]),
-            'acc_mag': np.array([1183.64226014, 1182.03933945, 1089.91788682, 991.98538296, 1152.54067173])
+        self.expected_units = {
+            'ECG': 'uV',
+            'ACC': 'mg',
+            'HRV': 'ms'
         }
-        self.expected_start_time = pd.Timestamp("3.1.2019 16:12:43")
+        self.expected_edf_metadata = [
+            {'label': 'ECG',
+            'dimension': 'uV',
+            'sample_rate': 500.0,
+            'physical_max': 32767.0,
+            'physical_min': -32768.0,
+            'digital_max': 32767,
+            'digital_min': -32768,
+            'prefilter': '',
+            'transducer': 'ECG electrode'},
+            {'label': 'Accelerometer_X',
+            'dimension': 'mg',
+            'sample_rate': 25.0,
+            'physical_max': 4000.0,
+            'physical_min': -4000.0,
+            'digital_max': 8000,
+            'digital_min': -8000,
+            'prefilter': '',
+            'transducer': 'X-axis'},
+            {'label': 'Accelerometer_Y',
+            'dimension': 'mg',
+            'sample_rate': 25.0,
+            'physical_max': 4000.0,
+            'physical_min': -4000.0,
[…]
