### 2021-07-18 ae8e1cb
add unit test for timeshifting faros
