+++ tests/test_faros.py
@@ -213,48 +213,80 @@ class FarosTestCase(unittest.TestCase):
                                           expected_edf_metadata=self.expected_edf_metadata)
 
         os.remove(self.EDF_WRITE_PATH)
-    
 
-    
+    def test_timeshift_to_timestamp(self):
+        timestamp = pd.Timestamp('1.2.2019, 21:43:23')
+
+        expected_shifted_ecg_index_head = pd.DatetimeIndex(['2019-01-02 21:43:23', '2019-01-02 21:43:23.002000',
+                                                            '2019-01-02 21:43:23.004000', '2019-01-02 21:43:23.006000',
+                                                            '2019-01-02 21:43:23.008000'])
+        expected_shifted_acc_index_head = pd.DatetimeIndex(['2019-01-02 21:43:23', '2019-01-02 21:43:23.040000',
+                                                            '2019-01-02 21:43:23.080000', '2019-01-02 21:43:23.120000',
+                                                            '2019-01-02 21:43:23.160000'])
+        expected_shifted_marker_index_head = pd.DatetimeIndex(['2019-01-02 21:43:23', '2019-01-02 21:43:24',
+                                                               '2019-01-02 21:43:25', '2019-01-02 21:43:26',
+                                                               '2019-01-02 21:43:27'])
+        expected_shifted_hrv_index_head = pd.DatetimeIndex(['2019-01-02 21:43:23', '2019-01-02 21:43:23.200000',
+                                                            '2019-01-02 21:43:23.400000', '2019-01-02 21:43:23.600000',
+                                                            '2019-01-02 21:43:23.800000'])
+        expected_shifted_data_index_head = pd.DatetimeIndex(['2019-01-02 21:43:23', '2019-01-02 21:43:23.002000',
+                                                            '2019-01-02 21:43:23.004000', '2019-01-02 21:43:23.006000',
+                                                            '2019-01-02 21:43:23.008000'])
+
+        # We need to use a reader other then self.reader_from_edf or self.reader_from_dir
+        # to not change any of their attributes.
+        reader = devicely.FarosReader(self.EDF_READ_PATH)
+        reader.timeshift(timestamp)
+
+        self.assertEqual(reader.start_time, timestamp)
+        pd.testing.assert_index_equal(reader.ECG.head().index, expected_shifted_ecg_index_head)
+        pd.testing.assert_index_equal(reader.ACC.head().index, expected_shifted_acc_index_head)
+        pd.testing.assert_index_equal(reader.Marker.head().index, expected_shifted_marker_index_head)
+        pd.testing.assert_index_equal(reader.HRV.head().index, expected_shifted_hrv_index_head)
+        pd.testing.assert_index_equal(reader.data.head().index, expected_shifted_data_index_head)
 
 
-    # def test_timeshift_by_timedelta(self):
-    #     reader = devicely.FarosReader(self.READ_PATH)
-    #     shift = pd.Timedelta('1 days, 2 hours, 3 minutes, 4 seconds, 5 milliseconds')
-    #     reader.timeshift(shift)
-    #     expected_shifted_index_head = pd.DatetimeIndex(['2019-03-02 18:15:47.005000', '2019-03-02 18:15:47.006000',
-    #                                                     '2019-03-02 18:15:47.007000', '2019-03-02 18:15:47.008000',
-    #                                                     '2019-03-02 18:15:47.009000'])
-    #     pd.testing.assert_index_equal(reader.data.head().index, expected_shifted_index_head, check_names=False)
-    #     self.assertEqual(reader.start_time, pd.Timestamp('2019-03-02 18:15:47.005000'))
+    def test_timeshift_by_timedelta(self):
[…]
