### 2021-07-19 a61aee3
adjust spacelabs set_window method
