### 2021-07-19 33b08d4
remove code duplication in spacelabs timeshift, changed some spacelabs tests
