+++ devicely/spacelabs.py
@@ -8,7 +8,6 @@ import random
 from xml.etree import ElementTree as ET
 
 import pandas as pd
-import xmltodict
 
 
 class SpacelabsReader:
@@ -55,10 +54,8 @@ class SpacelabsReader:
             metadata = pd.read_csv(path, nrows=6, header=None)
             self.valid_measurements = str(metadata.loc[5, 0])
 
-        column_names = ['hour', 'minutes', 'SYS(mmHg)', 'DIA(mmHg)', 'x', 'y', 'error', 'z']
-        self.data = pd.read_csv(path, sep=',', skiprows=51, skipfooter=1, header=None,
-                                names=column_names,
-                                engine='python')
+        column_names = ['hour', 'minutes', 'SYS(mmHg)', 'DIA(mmHg)', 'ACC_x', 'ACC_y', 'error', 'ACC_z']
+        self.data = pd.read_csv(path, sep=',', skiprows=51, skipfooter=1, header=None, names=column_names, engine='python')
 
         # Adjusting Date
         dates = [base_date]
@@ -74,7 +71,7 @@ class SpacelabsReader:
         self.data['date'] = dates
         self.data['time'] = times
 
-        order = ['timestamp', 'date', 'time', 'SYS(mmHg)', 'DIA(mmHg)', 'x', 'y', 'z', 'error']
+        order = ['timestamp', 'date', 'time', 'SYS(mmHg)', 'DIA(mmHg)', 'ACC_x', 'ACC_y', 'ACC_z', 'error']
         self.data = self.data[order]
 
         xml_line = open(path, 'r').readlines()[-1]
@@ -96,15 +93,15 @@ class SpacelabsReader:
         if subject_id:
             self.subject = subject_id
         else:
-            self.subject = 'xxxxxx'
+            self.subject = ''
 
         self.metadata = {
-            'PATIENTINFO' : {'DOB' : None,
-                             'RACE' : None},
-            'REPORTINFO' : {'PHYSICIAN' : None,
-                            'NURSETECH' : None,
-                            'STATUS' : None,
-                            'CALIPERSUMMARY' : {'COUNT' : None}}
+            'PATIENTINFO' : {'DOB' : '',
+                             'RACE' : ''},
+            'REPORTINFO' : {'PHYSICIAN' : '',
+                            'NURSETECH' : '',
+                            'STATUS' : '',
+                            'CALIPERSUMMARY' : {'COUNT' : ''}}
         }
 
     def write(self, path):
@@ -131,21 +128,21 @@ class SpacelabsReader:
             printing_df = self.data.drop(columns=['date', 'time'])
             printing_df['hours'] = self.data.time.map(lambda x: x.strftime("%H"))
             printing_df['minutes'] = self.data.time.map(lambda x: x.strftime("%M"))
-            order = ['hours', 'minutes', 'SYS(mmHg)', 'DIA(mmHg)', 'x', 'y', 'error', 'z']
+            order = ['hours', 'minutes', 'SYS(mmHg)', 'DIA(mmHg)', 'ACC_x', 'ACC_y', 'error', 'ACC_z']
             printing_df = printing_df[order]
             printing_df.fillna(-9999, inplace=True)
             printing_df.replace('EB', -9998, inplace=True)
             printing_df.replace('AB', -9997, inplace=True)
-            printing_df[['SYS(mmHg)', 'DIA(mmHg)', 'x', 'y', 'error', 'z']] = printing_df[
-                ['SYS(mmHg)', 'DIA(mmHg)', 'x', 'y', 'error', 'z']].astype(int).astype(str)
[…]
