### 2021-07-19 81150f9
faros test not ignored anymore in github action
