+++ .github/workflows/test.yml
@@ -30,12 +30,12 @@ jobs:
     
     - name: Test with pytest
       run: |
-         python -m pytest --ignore=tests/test_faros.py
+         python -m pytest
 
     - if: ${{ matrix.python-version == '3.9' }}
       name: Calculate coverage
       run: |
-         coverage run --source=tests -m pytest --ignore=tests/test_faros.py
+         coverage run --source=tests -m pytest
          coverage xml
          COVERAGE=$(python -c "import xml.etree.ElementTree as ET; print(int(float(ET.parse('coverage.xml').getroot().attrib['line-rate']) * 100))")
          COLOR=$(echo $COVERAGE | python -c "import sys; from bisect import bisect; i=bisect([0,60,70,80,95,100], int(sys.stdin.read()))-1; print(['red', 'orange', 'yellow', 'yellowgreen', 'green', 'brightgreen'][i])")
