### 2021-07-19 db2679d
change github action to calculate coverage correctly
