+++ tests/test_spacelabs.py
@@ -64,159 +64,159 @@ class SpacelabsTestCase(unittest.TestCase):
     def setUp(self):
         self.spacelabs_reader = devicely.SpacelabsReader(self.READ_PATH)
 
-    # def test_read(self):
-    #     # Tests if a basic reading operation.
+    def test_read(self):
+        # Tests if a basic reading operation.
 
-    #     pd.testing.assert_frame_equal(self.spacelabs_reader.data,
-    #                                   self.expected_data)
-    #     self.assertEqual(self.spacelabs_reader.subject, self.expected_subject)
-    #     self.assertEqual(self.spacelabs_reader.metadata,
-    #                      self.expected_metadata)
+        pd.testing.assert_frame_equal(self.spacelabs_reader.data,
+                                      self.expected_data)
+        self.assertEqual(self.spacelabs_reader.subject, self.expected_subject)
+        self.assertEqual(self.spacelabs_reader.metadata,
+                         self.expected_metadata)
 
-    # def test_deidentify(self):
-    #     # Tests if the SpacelabsReader.deidentify method removes all patient metadata.
+    def test_deidentify(self):
+        # Tests if the SpacelabsReader.deidentify method removes all patient metadata.
 
-    #     self.spacelabs_reader.deidentify()
+        self.spacelabs_reader.deidentify()
 
-    #     self.assertEqual(self.spacelabs_reader.subject, "")
-    #     self.assertEqual(
-    #         self.spacelabs_reader.metadata,
-    #         {
-    #             "PATIENTINFO": {
-    #                 "DOB": "",
-    #                 "RACE": ""
-    #             },
-    #             "REPORTINFO": {
-    #                 "PHYSICIAN": "",
-    #                 "NURSETECH": "",
-    #                 "STATUS": "",
-    #                 "CALIPERSUMMARY": {
-    #                     "COUNT": ""
-    #                 },
-    #             },
-    #         },
-    #     )
+        self.assertEqual(self.spacelabs_reader.subject, "")
+        self.assertEqual(
+            self.spacelabs_reader.metadata,
+            {
+                "PATIENTINFO": {
+                    "DOB": "",
+                    "RACE": ""
+                },
+                "REPORTINFO": {
+                    "PHYSICIAN": "",
+                    "NURSETECH": "",
+                    "STATUS": "",
+                    "CALIPERSUMMARY": {
+                        "COUNT": ""
+                    },
+                },
+            },
+        )
 
-    # def test_write(self):
-    #     self.spacelabs_reader.write(self.WRITE_PATH)
-    #     new_reader = devicely.SpacelabsReader(self.WRITE_PATH)
+    def test_write(self):
+        self.spacelabs_reader.write(self.WRITE_PATH)
+        new_reader = devicely.SpacelabsReader(self.WRITE_PATH)
 
-    #     pd.testing.assert_frame_equal(new_reader.data,
-    #                                   self.spacelabs_reader.data)
-    #     self.assertEqual(new_reader.metadata, self.spacelabs_reader.metadata)
-    #     self.assertEqual(new_reader.subject, self.spacelabs_reader.subject)
+        pd.testing.assert_frame_equal(new_reader.data,
+                                      self.spacelabs_reader.data)
+        self.assertEqual(new_reader.metadata, self.spacelabs_reader.metadata)
[…]
