### 2021-07-19 b178955
remove step "show Poetry version" in github workflow
