+++ .github/workflows/test.yml
@@ -24,10 +24,6 @@ jobs:
     
     - name: Install Poetry
       uses: snok/install-poetry@v1
-      
-    - name: show Poetry version
-      run: |
-        poetry -V
 
     - name: Install dependencies
       run: |
