### 2021-07-19 9172a98
improve docstrings in faros
