### 2021-07-19 0cb4e3c
add comment in test_spacelabs.py
