+++ CHANGELOG.md
@@ -1,0 +1,19 @@
+# Changelog
+All notable changes to this project will be documented in this file.
+
+The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
+and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+
+## [Unreleased]
+
+## [1.0.0] - 2021-07-19
+### Added
+- devicely.FarosReader can both read from and write to EDF files and directories
+- devicely.FarosReader has as attributes the individual dataframes (ACC, ECG, ...) and not only the joined dataframe
+
+### Changed
+- in devicely.SpacelabsReader, use xml.etree from the standard library instead of third-party "xmltodict"
+- switch from setuptools to Poetry
+
+### Removed
+- removed setup.py because static project files such as pyproject.toml are preferred
