### 2021-07-19 #37 @arianesasso
Hello @xmnlab! Some updates from our side :).

I think we covered most requests by @agricolab

- Added test coverage and repostatus.org badges
- Tests should not be failing anymore
- Faros is not being ignored anymore
- Created version 1.0.0 and added HISTORY/CHANGELOG.md (will have it in pypi this week)
- **Missing:** A comparison to other packages should be included in the paper

We also added some changes based on his and @willingc great reviews:

- Changed faros test file (because the size was too big) and fixed tests
- Made possible for people to read Faros back into EDF format (when there are no changes into the data frame)
- Changed setup.py to pyproject.toml together with poetry (for packaging; dependancy management and building for pypi)
- Now we can run tests with poetry as well: poetry install; poetry run python -m pytest
- Improved the test coverage

Also, from her side we covered:

- Badges
- Renamed master branch to main and update test.yml, etc.
- I think now most comments related to pipenv are depracated since we moved to poetry
- Improved test coverage

Now, we are working on the comments related to documentation. Since they had many great points we think we have to make it better, plan:

- Improve readme and documentation (with embedded jupyter cells)
- Use binder
- Improve installation, contributing and testing documentation
- Update the paper and add it to sphinx

We plan to give you a new update by Friday :).
