### 2021-07-19 #37 @xmnlab
@arianesasso , thank you so much for this detailed update!
The update sounds great! I am excited and looking forward to seeing your next update :)
