### 2021-07-24 0e72f66
Empatica: make compliant with interface of other readers, improved unit test code quality
