+++ devicely/empatica.py
@@ -57,9 +57,6 @@ class EmpaticaReader:
         across signals.
     """
 
-    signal_names = ['acc', 'bvp', 'eda', 'hr', 'temp']
-    acc_names = ['acc_x', 'acc_y', 'acc_z']
-
     def __init__(self, path):
         """
         Parse the csv files located in the specified directory into dataframes.
@@ -77,29 +74,23 @@ class EmpaticaReader:
             In case path is not specified.
         """
 
-        self._init_filelist(path)
         self.start_times = {}
         self.sample_freqs = {}
 
-        self.BVP = self._read_signal('bvp')
-        self.EDA = self._read_signal('eda')
-        self.HR = self._read_signal('hr')
-        self.TEMP = self._read_signal('temp')
-        self.ACC = self._read_acc()
-        self.IBI = self._read_ibi()
+        self.ACC = self._read_signal(os.path.join(path, 'ACC.csv'), 'ACC', col_names=['X', 'Y', 'Z'])
+        if self.ACC is not None:
+            self.ACC['mag'] = np.linalg.norm(self.ACC.values, axis=1)
+        self.BVP = self._read_signal(os.path.join(path, 'BVP.csv'), 'BVP')
+        self.EDA = self._read_signal(os.path.join(path, 'EDA.csv'), 'EDA')
+        self.HR = self._read_signal(os.path.join(path, 'HR.csv'), 'HR')
+        self.TEMP = self._read_signal(os.path.join(path, 'TEMP.csv'), 'TEMP')
+        self.IBI = self._read_ibi(os.path.join(path, 'IBI.csv'))
 
-        if not file_empty_or_not_existing(self.filelist['tags']):
-            self.tags = pd.read_csv(self.filelist['tags'], header=None, parse_dates=[0],
-                                    date_parser=lambda x: pd.Timestamp(float(x), unit='s'))
-        else:
-            self.tags = None
+        self.tags = self._read_tags(os.path.join(path, 'tags.csv'))
+        
+        self.data = self._get_joined_dataframe()
 
-        if ((self.BVP is None) & (self.EDA is None) & (self.HR is None)
-            & (self.TEMP is None) & (self.ACC is None) & (self.IBI is None)):
-            raise Exception('All Files are Empty.')
-        self._update_joined_dataframe()
-
-    def write(self, path):
+    def write(self, dir_path):
         """
         Write the signal dataframes back to individual csv files formatted the
         same way as the read csv files.
@@ -120,88 +111,75 @@ class EmpaticaReader:
             In case path is not specified.
         """
 
-        if not path:
-            raise Exception('Please specify a path to save the files.')
-        if not os.path.exists(path):
-            os.mkdir(path)
-        self._write_signal(path, self.BVP)
-        self._write_signal(path, self.EDA)
-        self._write_signal(path, self.HR)
-        self._write_signal(path, self.TEMP)
-        self._write_acc(path)
-        self._write_ibi(path)
+        if not os.path.exists(dir_path):
+            os.mkdir(dir_path)
+        self._write_signal(os.path.join(dir_path, 'ACC.csv'), self.ACC.drop(columns='mag'), 'ACC')
+        self._write_signal(os.path.join(dir_path, 'BVP.csv'), self.BVP, 'BVP')
+        self._write_signal(os.path.join(dir_path, 'EDA.csv'), self.EDA, 'EDA')
+        self._write_signal(os.path.join(dir_path, 'HR.csv'), self.HR, 'HR')
+        self._write_signal(os.path.join(dir_path, 'TEMP.csv'), self.TEMP, 'TEMP')
+        self._write_ibi(os.path.join(dir_path, 'IBI.csv'))
[…]
