+++ tests/test_empatica.py
@@ -10,165 +10,225 @@ class EmpaticaTestCase(unittest.TestCase):
     READ_PATH = 'tests/Empatica_test_data/test_data_read'
     WRITE_PATH = 'tests/Empatica_test_data/test_data_write'
 
-    def setUp(self):
-        self.empatica_reader = devicely.EmpaticaReader(self.READ_PATH)
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
 
-        self.start_times = {
-            'acc': pd.Timestamp(1551453301, unit='s'),
-            'bvp': pd.Timestamp(1551453301, unit='s'),
-            'eda': pd.Timestamp(1551453301, unit='s'),
-            'hr': pd.Timestamp(1551453311, unit='s'),
-            'ibi': pd.Timestamp(1551453301, unit='s'),
-            'temp': pd.Timestamp(1551453301, unit='s'),
+        self.reader = devicely.EmpaticaReader(self.READ_PATH)
+
+        self.expected_start_times = {
+            'ACC': pd.Timestamp(1551453301, unit='s'),
+            'BVP': pd.Timestamp(1551453301, unit='s'),
+            'EDA': pd.Timestamp(1551453301, unit='s'),
+            'HR': pd.Timestamp(1551453311, unit='s'),
+            'IBI': pd.Timestamp(1551453301, unit='s'),
+            'TEMP': pd.Timestamp(1551453301, unit='s'),
         }
-        self.sample_freqs = {
-            'acc': 32,
-            'bvp': 64,
-            'eda': 4,
-            'hr': 1,
-            'temp': 4,
+
+        self.expected_sample_freqs = {
+            'ACC': 32,
+            'BVP': 64,
+            'EDA': 4,
+            'HR': 1,
+            'TEMP': 4,
         }
-        self.acc_df = pd.DataFrame({
-            'acc_x': [-1.0, -1.0, -1.0, -1.0, -1.0, -1.0, -1.0, -1.0, -1.0, -1.0, -1.0, -1.0, 0.0, 0.0, -1.0, -1.0,
-                      -1.0, -1.0, -1.0, -1.0],
-            'acc_y': [65.0, 65.0, 65.0, 65.0, 64.0, 64.0, 65.0, 65.0, 65.0, 65.0, 65.0, 64.0, 65.0, 65.0, 65.0,
-                      64.0, 64.0, 64.0, 65.0, 65.0],
-            'acc_z': [5.0, 5.0, 4.0, 5.0, 5.0, 5.0, 5.0, 4.0, 5.0, 5.0, 5.0, 5.0, 5.0, 5.0, 5.0, 5.0, 4.0, 5.0, 5.0,
-                      5.0],
-            'acc_mag': [65.19969325081216, 65.19969325081216, 65.13063795173512, 65.19969325081216,
-                        64.20280367709809, 64.20280367709809, 65.19969325081216, 65.13063795173512,
-                        65.19969325081216, 65.19969325081216, 65.19969325081216, 64.20280367709809,
-                        65.19202405202648, 65.19202405202648, 65.19969325081216, 64.20280367709809,
-                        64.13267497929586, 64.20280367709809, 65.19969325081216, 65.19969325081216]
-        })
-        self.acc_df.index = pd.to_datetime(['2019-03-01 15:15:01', '2019-03-01 15:15:01.031250',
-                                           '2019-03-01 15:15:01.062500', '2019-03-01 15:15:01.093750',
-                                           '2019-03-01 15:15:01.125000', '2019-03-01 15:15:01.156250',
-                                           '2019-03-01 15:15:01.187500', '2019-03-01 15:15:01.218750',
-                                           '2019-03-01 15:15:01.250000', '2019-03-01 15:15:01.281250',
[…]
