### 2021-07-24 1deeb32
slight modifications to docstrings in empatica
