### 2021-07-25 #37 @agricolab
Great. I'm in vacation until start of August. Will review next week, but I guess most stuff is covered.
