### 2021-07-25 8120ded
edit example.ipynb, shimmer refactoring
