+++ devicely/shimmer_plus.py
@@ -1,7 +1,9 @@
 """
 Module to process Shimmer Plus data
 """
+from os import sep
 import random
+import re
 
 import numpy as np
 import pandas as pd
@@ -15,12 +17,12 @@ class ShimmerPlusReader:
     Attributes
     ----------
     data : DataFrame
-        DataFrame with the values that were read from the csv file
+        DataFrame with the read signals. The column 'Shimmer_40AC_Timestamp_Unix_CAL' contains the timestamps of measurement.
 
     units : pandas.Series
         a unit for each column in the dataframe
     """
-    def __init__(self, path, delimiter=';'):
+    def __init__(self, path):
         """
         Reads the csv file generated by the Shimmer device and saves the
         parsed DataFrame.
@@ -32,23 +34,20 @@ class ShimmerPlusReader:
         delimiter, optional : str
             delimiter used in the csv file
         """
-        self.delimiter = delimiter
-        self.data = pd.read_csv(path, sep=delimiter, skiprows=1)
-        self.data = self.data.dropna(axis=1, how='all')
-        self.units = self.data.iloc[0] # saving self.data units
-        self.data = self.data.drop(0).astype(float)
-        self.data['time'] = pd.to_datetime(
-            self.data['Shimmer_40AC_Timestamp_Unix_CAL'],
-            unit=self.units['Shimmer_40AC_Timestamp_Unix_CAL']).round('ms')
-        self.data.drop(columns=['Shimmer_40AC_Timestamp_Unix_CAL'], inplace=True)
-        self.data = self.data.drop_duplicates(subset='time')
-        self.data.set_index('time', inplace=True, verify_integrity=True)
-        self.data.sort_index(inplace=True)
-        self.data['acc_mag'] = np.linalg.norm(self.data[['Shimmer_40AC_Accel_LN_X_CAL',
-                                                         'Shimmer_40AC_Accel_LN_Y_CAL',
-                                                         'Shimmer_40AC_Accel_LN_Z_CAL']],
-                                              axis=1)
-        self.units['acc_mag'] = self.units['Shimmer_40AC_Accel_LN_Z_CAL']
+        with open(path, 'r') as f:
+            self.delimiter = re.findall(f'"sep=(.)"', f.readline())[0]
+            col_names = list(filter(bool, f.readline().strip().split(self.delimiter)))
+            units = list(filter(bool, f.readline().strip().split(self.delimiter)))
+            self.units = dict(zip(col_names, units))
+            self.data = pd.read_csv(f, sep=self.delimiter, header=None, index_col=False,
+                                                           names=col_names,
+                                                           parse_dates=['Shimmer_40AC_Timestamp_Unix_CAL'],
+                                                           date_parser=lambda x: pd.to_datetime(x, unit='ms').round('ms')
+                                                           )
+        self.data['Shimmer_40AC_Accel_LN_mag'] = np.linalg.norm(self.data[['Shimmer_40AC_Accel_LN_X_CAL',
+                                                                           'Shimmer_40AC_Accel_LN_Y_CAL',
+                                                                           'Shimmer_40AC_Accel_LN_Z_CAL']],
+                                                                axis=1)
 
     def write(self, path):
         """
@@ -60,13 +59,14 @@ class ShimmerPlusReader:
         path : str
             Path to writing csv file. Writing mode: 'w'.
         """
-        write_df = self.data.drop(columns=['acc_mag'])
[…]
