+++ pyproject.toml
@@ -29,6 +29,8 @@ pytest = "^6.2.4"
 coverage = "^5.5"
 ipykernel = "^6.0.3"
 jupyter = "^1.0.0"
+nbsphinx = "^0.8.6"
+insegel = "^1.1.0"
 
 [build-system]
 requires = ["poetry-core>=1.0.0"]
