+++ devicely/everion.py
@@ -1,6 +1,3 @@
-"""
-Module to process Everion data
-"""
 import os
 import glob
 import random
@@ -12,7 +9,7 @@ class EverionReader:
     Parses, timeshifts and writes data generated by Biovotion Everion. You can find example data `here <https://github.com/jostmorgenstern/devicely-documentation-sample-data/tree/main/Everion/>`_.
     Each measurement has a tag which is an information about the type of measurement. By default, only some of the tags are read for signals, sensors and features.
     You can find these tags via self.default_signal_tags, self.default_sensor_tags and self.default_feature_tags.
-    You can find out the meaning of all tags with self.SIGNAL_TAGS, self.SENSOR_TAGS and self.FEATURE_TAGS.
+    You can find out all available tags and what they mean by accessing the attributes EverionReader.SIGNAL_TAGS, EverionReader.SENSOR_TAGS and EverionReader.FEATURE_TAGS.
 
 
     Attributes
