### 2021-07-26 9ad1b98
move new docs html to docs directory
