+++ docs/_sources/index.rst.txt
@@ -1,12 +1,7 @@
-.. devicely documentation master file, created by
-   sphinx-quickstart on Wed Nov 25 00:15:47 2020.
-   You can adapt this file completely to your liking, but it should at least
-   contain the root `toctree` directive.
+devicely, a Python package for accessing health sensor data
+===========================================================
 
-Introduction
-============
-
-The devicely package is made for reading, writing and timeshifting health sensor data from:
+The devicely package is made for reading, writing and anonymizing health sensor data from several sensors:
 
 * `Empatica E4 (Firmware 2.1.0.4911) <https://e4.empatica.com/e4-wristband/>`_
 * `Biovotion Everion (Firmware 03.06) <https://www.biovotion.com/everion/>`_
@@ -16,352 +11,35 @@ The devicely package is made for reading, writing and timeshifting health sensor
 * `Shimmer Consensys GSR (Shimmer3 GSR Development Kit) <https://www.shimmersensing.com/products/gsr-optical-pulse-development-kit#specifications-tab/>`_
 
 
-Examples
-========
+devicely makes it easy to read sensor data and access it in common formats such as dataframes.
 
-Empatica
-########
+We provice a custom reader class for each sensor which provides three core methods: read, timeshift and anonymize.
+The timeshift method changes the time of measurement, thereby automatically anonymizing the data.
 
-The Empatice E4 wristband can be used to obtain data from inter-beat intervals, electrodermal activity, heart rate, temperature and blood volume pulse.
-The wristband outputs its measurements in `this directory structure <https://github.com/jostmorgenstern/devicely-documentation-sample-data/tree/main/Empatica/>`_.
+After timeshifting, you can write the data in the same original sensor format.
+Now you are safe to distribute the data with others.
 
-Create an :code:`EmpaticaReader` object, passing as a parameter a path of the same directory structure:
 
-.. code-block:: Python
+Usage
+-----
 
-   >>> import devicely
-   >>> empatica_reader = devicely.EmpaticaReader('path/to/empatica/dir')
+- :doc:`example`
+- :doc:`moduleref`
 
-Accessing the individual signal dataframes:
 
-.. code-block:: Python
+About devicely
+--------------
 
-   >>> empatica_reader.ACC.head(3)
-                              acc_x  acc_y  acc_z    acc_mag
-   2020-10-16 17:04:29.000000     -3     11     60  61.073726
-   2020-10-16 17:04:29.031250     -3     11     60  61.073726
-   2020-10-16 17:04:29.062500     -2     11     60  61.032778
+- `GitHub <https://github.com/hpi-dhc/devicely>`_
+- `PyPi <https://pypi.org/project/devicely/>`_
+- :doc:`contribution`
 
-You can also access :code:`empatica_reader.IBI` (inter-beat intervals), :code:`empatica_reader.EDA` (electrodermal activity), :code:`empatica_reader.HR` (heart rate), :code:`empatica_reader.TEMP` (temperature) and :code:`empatica_reader.BVP` (blood volume pulse).
-
-An EmpaticaReader object also contains a dataframe of all signals, joined by time of sampling. Because of varying sampling frequencies across signals, expect to see a lot of nan values.
-
-.. code-block:: Python
-
-   >>> empatica_reader.data.head(3)
-                              bvp  eda  hr   temp  acc_x  acc_y  acc_z    acc_mag  timedelta  ibi
-   2020-10-16 17:04:29.000000 -0.0  0.0 NaN  31.13   -3.0   11.0   60.0  61.073726        NaN  NaN
[…]
