### 2021-07-26 8920775
add readme tag to pyproject.toml
