### 2021-08-03 4b736ff
Added related work section in the paper, updated author list and contributors, improved the text overall
