+++ inst/paper.bib
@@ -48,3 +48,63 @@ url = {https://doi.org/10.1038/s41551-020-00640-6},
 volume = {4},
 year = {2020}
 }
+
+@article{Snyder2020,
+  doi = {10.21105/joss.02106},
+  url = {https://doi.org/10.21105/joss.02106},
+  year = {2020},
+  publisher = {The Open Journal},
+  volume = {5},
+  number = {47},
+  pages = {2106},
+  author = {Phil Snyder and Meghasyam Tummalacherla and Thanneer Perumal and Larsson Omberg},
+  title = {mhealthtools: A Modular R Package for Extracting Features from Mobile and Wearable Sensor Data},
+  journal = {Journal of Open Source Software}
+}
+
+@article{Christakis2019,
+  doi = {10.21105/joss.01663},
+  url = {https://doi.org/10.21105/joss.01663},
+  year = {2019},
+  publisher = {The Open Journal},
+  volume = {4},
+  number = {44},
+  pages = {1663},
+  author = {Yiorgos Christakis and Nikhil Mahadevan and Shyamal Patel},
+  title = {SleepPy: A python package for sleep analysis from accelerometer data},
+  journal = {Journal of Open Source Software}
+}
+
+@article{Bartels2020,
+  doi = {10.21105/joss.01867},
+  url = {https://doi.org/10.21105/joss.01867},
+  year = {2020},
+  publisher = {The Open Journal},
+  volume = {5},
+  number = {51},
+  pages = {1867},
+  author = {Rhenan Bartels and Tiago Peçanha},
+  title = {HRV: a Pythonic package for Heart Rate Variability Analysis},
+  journal = {Journal of Open Source Software}
+}
+
+@article{Czech2019,
+  doi = {10.21105/joss.01778},
+  url = {https://doi.org/10.21105/joss.01778},
+  year = {2019},
+  publisher = {The Open Journal},
+  volume = {4},
+  number = {43},
+  pages = {1778},
+  author = {Matthew D. Czech and Shyamal Patel},
+  title = {GaitPy: An Open-Source Python Package for Gait Analysis Using an Accelerometer on the Lower Back},
+  journal = {Journal of Open Source Software}
+}
+
+@misc{Maritsch2020,
+author = {Maritsch, Martin and F{\"{o}}ll, Simon and Fleisch, Elgar and Kowatsch, Tobias and Wortmann, Felix and Spinola, Federica},
+booktitle = {Chair of Information Management at ETH Zurich and the Institute of Technology Management at University of St. Gallen},
+title = {{FLIRT: Feature generation tooLkIt for weaRable daTa}},
+url = {https://github.com/im-ethz/flirt},
+year = {2020}
+}
