+++ inst/paper.md
@@ -14,6 +14,9 @@ authors:
 - name: Felix Musmann^[felix.musmann@student.hpi.uni-potsdam.de]
   orcid: 0000-0001-5365-0785
   affiliation: 1
+- name: Bert Arnrich^[bert.arnrich@hpi.de]
+  orcid: 0000-0001-5365-0785
+  affiliation: 1
 affiliations:
  - name: Digital Health Center, Hasso Plattner Institute, University of Potsdam
    index: 1
@@ -29,18 +32,49 @@ temperature, blood oxygen saturation, acceleration, blood glucose and much more
 increase in market presence in 2020 alone `[@IDC2020:2020]`. Applications for wearable
 devices vary from tracking cardiovascular risks `[@Bayoumy2021:2021]` to identifying
 COVID-19 onset `[@Mishra2020:2020]`. Therefore, there is a great need for scientists to
-easily go through data acquired from different wearables.
+easily go through data acquired from different wearables and to be able to share
+them while protecting patient privacy.
 In order to solve this problem and empower scientists working with biosignals,
 we developed the **devicely** package. It represents the data in a science-friendly
-format and lets scientists focus on what they want: the analysis of biosignals.
+format and contains tools for data deidentification. It allows scientists to focus
+on what they want: the analysis of biosignals guided by privacy principles.
+
+# Related Work
+
+The first example of a package working with wearable data is mhealthtools `[@Snyder2020:2020]`, which is developed in R and focus on
+extracting features from sensors such as inertial measurement units (IMUs). Its main difference from **devicely**
+is firstly the language (R versus Python) and secondly their complementary
+nature. Mhealthtools offers functionalities for feature extraction and
+**devicely** is
+intended to help users a step before by reading and writing data from
+wearables into standardized formats.
+
+There are also packages developed in python, such as SleepPy `[@Christakis2019:2019]` which uses raw
+accelerometer data for assessing sleep quantity and quality. The HRV
+ `[@Bartels2020:2020]` package uses csv and text files or python iterables such as lists to generate
+features related to heart rate variability (HRV). GaitPy `[@Czech2019:2019]` accepts input data in a
+customizable format and is mainly used to extract features for gait analysis.
+Therefore, packages such as SleepPy, HRV and GaitPy could also be used as a
+second step to extract features from the output generated by **devicely**.
+
+FLIRT `[@Maritsch2020:2020]` is the only package that provides ways for
+reading data from specific wearables such as Empatica E4. It also
+includes functionalities to extract features from electrodermal activity (EDA),
+acceleration and HRV. The main difference from FLIRT to **devicely** is the
+focus on feature extraction versus privacy and data sharing.
+FLIRT reads the data for extracting features, while **devicely** aims to provide users with
+a way to read the data, deidentify them as necessary and write them back. In this
+way, researchers can protect even more data privacy and use the data easily for
+further analysis and sharing.
+
 
 # Statement of need
 
-Every wearable company has a different data format and reading this data is
+Every wearable company has a different data format and reading them is
 usually a challenge for scientists. Therefore, in order for researchers to be able to use different sensor data in an easy and
-friendly way we developed the **devicely** package. The package also contains two methods to help with data _deidentification_, one
+friendly way we developed the **devicely** package. The package also contains two methods to help with data deidentification, one
 is called timeshift and the other is a write method. The idea behind them is
-that you can timeshift all your time series to a different time from the one the
+that researchers can timeshift all their time series data to a different time from the one the
 actual experiments occurred and then write this new deidentified dataset back to
 the original or a similar data format. This will empower scientists to keep patient privacy
 and hopefully share more data to increase research reproducibility.
[…]
