### 2021-08-03 #37 @arianesasso
@agricolab Thank you :). I updated the paper now with the related work section and some corrections.

We also put a lot of effort in improving the documentation and are still working on that. We are also missing the **conda-forge feedstock** but will try to have that as well by the end of the week. I think it is coming to an end! And then we will be done with both revisions.

Thank you for your patience!
