### 2021-08-11 faa5cea
docs contribution section: add how to use poetry/ run tests; remove insegel dependency from pyproject.toml
