+++ docs/_sources/installation.rst.txt
@@ -1,4 +1,0 @@
-How to install bla bla bla
-===========================
-
-bli bla blubb
