### 2021-08-13 e3e35ce
Removed acc mag, added flexibility to files read
