+++ devicely/empatica.py
@@ -1,3 +1,9 @@
+"""
+Empatica E4 is a wearable device that offers real-time physiological data
+acquisition such as blood volume pulse, electrodermal activity (EDA), heart
+rate, interbeat intervals, 3-axis acceleration and skin temperature.
+"""
+
 import os
 import random
 
@@ -7,45 +13,48 @@ import pandas as pd
 
 class EmpaticaReader:
     """
-    Parses, timeshifts and writes data generated by Empatica E4.
+    Read, timeshift and write data generated by Empatica E4.
 
     Attributes
     ----------
     start_times : dict
-        Contains the timestamp of the first measurement for all
-        measured signals (BVP, ACC etc.).
+        Contain the timestamp of the first measurement for all
+        measured signals (BVP, ACC, etc.).
 
-    sample_freqs : dict
-        Contains the sampling frequencies of all
-        measured signals in Hz.
+    sample_freqs : dict ]
+        Contain the sampling frequencies of all measured signals
+        in Hz.
 
     IBI : pandas.DataFrame
-        Contains inter-beat interval data.
-        The column "seconds_since_start" is the time in seconds between the start of measurements
-        and the column "IBI" is the duration in seconds between consecutive beats.
-        
+        Contain inter-beat interval data. The column
+        "seconds_since_start" is the time in seconds between the start of
+        measurements and the column "IBI" is the duration in seconds between
+        consecutive beats.
+
     ACC : pandas.DataFrame
-        Contains the data measured with the onboard MEMS type 3-axis
-        accelerometer, indexed by time of measurement.
+        Contain the data measured with the onboard MEMS type
+        3-axis accelerometer, indexed by time of measurement.
 
     BVP : pandas.DataFrame
-        Contains blood volume pulse data, indexed by
-        time of measurement.
+        Contain blood volume pulse data, indexed by time of
+        measurement.
 
     EDA : pandas.DataFrame
-        Contains data captured from the electrodermal activity
+        Contain data captured from the electrodermal activity
         sensor, indexed by time of measurement.
 
     HR : pandas.DataFrame
-        Contains heart rate data, indexed by time of measurement.
+        Contain heart rate data, indexed by time of
+        measurement.
 
     TEMP : pandas.DataFrame
-        Contains temperature data, indexed by time of measurement.
+        Contain temperature data, indexed by time of
+        measurement.
 
     data : pandas.DataFrame
-        Joined dataframe of the ACC, BVP, EDA, HR and TEMP dataframes (see
-        above). May contain NaN values because sampling frequencies differ
-        across signals.
+        Joined dataframe of the ACC, BVP, EDA, HR and TEMP
+        dataframes (see above). May contain NaN values because sampling
+        frequencies differ across signals.
     """
 
     def __init__(self, path):
@@ -57,15 +66,19 @@ class EmpaticaReader:
         path : str
             Path of the directory that contains the individual signal csv
             files. The files must be named ACC.csv, BVP.csv, EDA.csv, HR.csv,
-            IBI.csv and TEMP.csv. If present, the file tags.csv is read too.
+            IBI.csv and TEMP.csv. If present, the file tags.csv is also read.
         """
 
         self.start_times = {}
         self.sample_freqs = {}
 
+        files = [f for f in os.listdir(path) if os.path.isfile(os.path.join(path, f))]
+
+        if files is None:
+            print('Empty directory. Nothing to read.')
+            return None
+
[…]
