+++ devicely/everion.py
@@ -1,3 +1,11 @@
+""""
+Biovotion Everion is a wearable device used for the continuous monitoring of
+vital signs. Currently, it measures the following vital signs: heart rate, blood
+pulse wave, heart rate variability, activity, SPO2, blood perfusion, respiration
+rate, steps, energy expenditure, skin temperature, EDA / galvanic skin response
+(GSR), barometric pressure and sleep.
+"""
+
 import os
 import glob
 import random
@@ -6,38 +14,50 @@ import pandas as pd
 
 class EverionReader:
     """
-    Parses, timeshifts and writes data generated by Biovotion Everion. You can find example data `here <https://github.com/jostmorgenstern/devicely-documentation-sample-data/tree/main/Everion/>`_.
-    Each measurement has a tag which is an information about the type of measurement. By default, only some of the tags are read for signals, sensors and features.
-    You can find these tags via self.default_signal_tags, self.default_sensor_tags and self.default_feature_tags.
-    You can find out all available tags and what they mean by accessing the attributes EverionReader.SIGNAL_TAGS, EverionReader.SENSOR_TAGS and EverionReader.FEATURE_TAGS.
+    Read, timeshift and write data generated by Biovotion Everion.
+    You can find example data `here <https://github.com/jostmorgenstern/devicely-documentation-sample-data/tree/main/Everion/>`_.
+    Each measurement has a tag which is an information about the type of
+    measurement. By default, only some of the tags are read for signals, sensors
+    and features. You can find these tags via self.default_signal_tags,
+    self.default_sensor_tags and self.default_feature_tags. You can find out all
+    available tags and what they mean by accessing the attributes
+    EverionReader.SIGNAL_TAGS, EverionReader.SENSOR_TAGS and
+    EverionReader.FEATURE_TAGS.
 
 
     Attributes
     ----------
     raw dataframes : dataframes
-        These dataframes are accessible via the attributes aggregates, analytics_events, attributes_dailys, everion_events, features, sensors, signals.
-        Each dataframe corresponds to a file in the reading path.
+        These dataframes are accessible via the
+        attributes aggregates, analytics_events, attributes_dailys,
+        everion_events, features, sensors and signals. Each dataframe corresponds
+        to a file in the reading path.
 
     data : DataFrame
-        joined version of the raw dataframes (aggregates, analytics_events, attributes_dailys, everion_events, features, sensors, signals)
+        Joined version of the raw dataframes (aggregates,
+        analytics_events, attributes_dailys, everion_events, features, sensors and
+        signals).
 
     SIGNAL_TAGS : dict
-        signal tag numbers and their meaning
+        Signal tag numbers and their meaning.
 
     SENSOR_TAGS : dict
-        sensor tag numbers and their meaning
+        Sensor tag numbers and their meaning.
 
     FEATURE_TAGS : dict
-        feature tag numbers and their meaning
+        Feature tag numbers and their meaning.
 
     default_signal_tags : list
-        subset of tags that are read by default for signals
+        Subset of tags that are read by default for
+        signals.
 
     default_sensor_tags : list
-        subset of tags that are read by default for sensors
+        Subset of tags that are read by default for
+        sensors.
 
     default_feature_tags : list
-        subset of tags that are read by default for features
+        Subset of tags that are read by default for
+        features.
 
     """
     SIGNAL_TAGS = {
@@ -104,31 +124,37 @@ class EverionReader:
 
     ACC_NAMES = ['accx_data', 'accy_data', 'accz_data']
 
-    def __init__(self, path, signal_tags=default_signal_tags, sensor_tags=default_sensor_tags, feature_tags=default_feature_tags):
[…]
