+++ devicely/faros.py
@@ -1,3 +1,8 @@
+"""
+1-lead ECG monitor FarosTM 180 from Bittium is a one channel ECG monitor with
+sampling frequency up to 1000 Hz and a 3D acceleration sampling up to 100Hz.
+"""
+
 import json
 import os
 import random
@@ -9,45 +14,45 @@ import pyedflib as edf
 
 class FarosReader:
     """
-    Parses, timeshifts and writes data
+    Read, timeshift and write data
     generated by Bittium Faros devices.
 
     Attributes
     ----------
     start_time : pandas.Timestamp
-        start time of all measurements
-    
+        Start time of all measurements.
+
     sample_freqs : dict
-        sampling frequencies of all signals in Hz
+        Sampling frequencies of all signals in Hz.
 
     units : dict
-        units of all signals
+        Units of all signals
 
     ECG : pandas.Series
-        ECG signal, indexed by timestamp
-    
+        ECG signal, indexed by timestamp.
+
     ACC : pandas.DataFrame
-        three ACC axes, indexed by timestamp
+        Three ACC axes, indexed by timestamp.
 
     Marker : pandas.Series
-        Markers, indexed by timestamp
-    
+        Markers, indexed by timestamp.
+
     HRV : pandas.Series
-        HRV signal, indexed by timestamp
+        HRV signal, indexed by timestamp.
 
     data : DataFrame
-        contains all signals (ECG, ACC, Marker, HRV) indexed by timestamp
-        Because the signals have different sampling frequencies, a lot of values are nan.
+        Contain all signals (ECG, ACC, Marker, HRV) indexed by timestamp.
+        Since the signals have different sampling frequencies, many values will be NaN.
     """
 
     def __init__(self, path):
         """
-        Reads a Faros-generated EDF file or a directory created by a FarosReader.
+        Read a Faros-generated EDF file or a directory created by a FarosReader.
 
         Parameters
         ----------
         path : str
-            Can either be Faros-generated EDF file or a directory created by the FarosReader.write() method.
+            Can either be a Faros-generated EDF file or a directory created by the FarosReader.write() method.
         """
 
         self.start_time = None
@@ -98,8 +103,7 @@ class FarosReader:
                                                          periods=len(acc),
                                                          freq=f"{1/self.sample_freqs['ACC']}S")
         )
-        self.ACC['mag'] = np.linalg.norm(self.ACC, axis=1)
-        
+
         marker = reader.readSignal(4)
         self.Marker = pd.Series(marker, name='Marker',
                                 index=pd.date_range(start=self.start_time,
@@ -126,16 +130,15 @@ class FarosReader:
 
         self.ECG = self._read_dir_csv(os.path.join(path, 'ECG.csv'), self.start_time, self.sample_freqs['ECG'])
         self.ACC = self._read_dir_csv(os.path.join(path, 'ACC.csv'), self.start_time, self.sample_freqs['ACC'])
-        self.ACC['mag'] = np.linalg.norm(self.ACC, axis=1)
         self.Marker = self._read_dir_csv(os.path.join(path, 'Marker.csv'), self.start_time, self.sample_freqs['Marker'])
         self.HRV = self._read_dir_csv(os.path.join(path, 'HRV.csv'), self.start_time, self.sample_freqs['HRV'])
 
     def _read_dir_csv(self, path, start_time, sample_freq):
-        df = pd.read_csv(path)
-        idx = pd.date_range(start=start_time, periods=len(df), freq=f"{1/sample_freq}S")
-        df.index = idx
+        dataframe = pd.read_csv(path)
+        idx = pd.date_range(start=start_time, periods=len(dataframe), freq=f"{1/sample_freq}S")
+        dataframe.index = idx
 
-        return df.squeeze()
+        return dataframe.squeeze()
 
[…]
