+++ devicely/shimmer_plus.py
@@ -1,3 +1,10 @@
+"""
+Shimmer Consensys GSR is a device that is used to collect sensor data in real
+time and it contains sensors such as GSR / EDA, photoplethysmography (PPG),
+3-axis accelerometer, 3-axis gyroscope, 3-axis magnetometer & integrated
+altimeter.
+"""
+
 import random
 import re
 
@@ -7,7 +14,7 @@ import pandas as pd
 
 class ShimmerPlusReader:
     """
-    Parses, deidentifies (timeshifts) and writes data
+    Read, timeshift and write data
     generated by Shimmer Consensys GSR.
 
     Attributes
@@ -16,11 +23,11 @@ class ShimmerPlusReader:
         DataFrame with the read signals. The column 'Shimmer_40AC_Timestamp_Unix_CAL' contains the timestamps of measurement.
 
     units : pandas.Series
-        a unit for each column in the dataframe
+        A unit for each column in the dataframe.
     """
     def __init__(self, path):
         """
-        Reads the csv file generated by the Shimmer device and saves the
+        Read the csv file generated by the Shimmer device and saves the
         parsed DataFrame.
 
         Parameters
@@ -28,7 +35,7 @@ class ShimmerPlusReader:
         path : str
             Path of the csv file.
         delimiter, optional : str
-            delimiter used in the csv file
+            Delimiter used in the csv file.
         """
         with open(path, 'r') as f:
             self.delimiter = re.findall(f'"sep=(.)"', f.readline())[0]
@@ -47,7 +54,7 @@ class ShimmerPlusReader:
 
     def write(self, path):
         """
-        Writes the DataFrame and units to the
+        Write the DataFrame and units to the
         writing path in the same format as it was read.
 
         Parameters
@@ -66,7 +73,7 @@ class ShimmerPlusReader:
 
     def timeshift(self, shift='random'):
         """
-        Timeshifts the data by shifting all time related columns.
+        Timeshift the data by shifting all time related columns.
 
         Parameters
         ----------
@@ -77,7 +84,8 @@ class ShimmerPlusReader:
             If shift is a timdelta, shifts the data by that timedelta.
 
             If shift is a timestamp, shifts the data such that the earliest entry
-            is at that timestamp and the remaining values keep the same time distance to the first entry.
+            has that timestamp. The remaining values will mantain the same
+            time difference to the first entry.
         """
         if shift == 'random':
             one_month = pd.Timedelta('30 days').value
