### 2021-08-13 232e939
deprecated methods, fixed columns and timeindex
