+++ devicely/spacelabs.py
@@ -1,3 +1,8 @@
+"""
+Spacelabs is an oscillometric blood pressure (BP) monitor which can be used to
+automatically track a person's BP in specificed time intervals.
+
+"""
 import csv
 import datetime as dt
 import random
@@ -8,7 +13,7 @@ import pandas as pd
 
 class SpacelabsReader:
     """
-    Parses, timeshifts, deidentifies and writes data
+    Read, timeshift, deidentify and write data
     generated by Spacelabs(SL90217).
 
     Attributes
@@ -17,11 +22,11 @@ class SpacelabsReader:
         DataFrame with the values that were read from the abp file.
 
     subject : str
-        Contains the sujbect's id. Can be changed for
+        Contain the sujbect's id. Can be changed for
         deidentification.
 
     valid_measurements : str
-        Contains the number of valid measurements in the
+        Contain the number of valid measurements in the
         abp file.
 
     metadata : dict
@@ -50,7 +55,7 @@ class SpacelabsReader:
             metadata = pd.read_csv(path, nrows=6, header=None)
             self.valid_measurements = str(metadata.loc[5, 0])
 
-        column_names = ['hour', 'minutes', 'SYS(mmHg)', 'DIA(mmHg)', 'ACC_x', 'ACC_y', 'error', 'ACC_z']
+        column_names = ['hour', 'minutes', 'SYS(mmHg)', 'DIA(mmHg)', 'UNKNOW_1', 'UNKNOW_2', 'CODE', 'UNKNOW_3']
         self.data = pd.read_csv(path, sep=',', skiprows=51, skipfooter=1, header=None, names=column_names, engine='python')
 
         # Adjusting Date
@@ -67,13 +72,14 @@ class SpacelabsReader:
         self.data['date'] = dates
         self.data['time'] = times
 
-        order = ['timestamp', 'date', 'time', 'SYS(mmHg)', 'DIA(mmHg)', 'ACC_x', 'ACC_y', 'ACC_z', 'error']
+        order = ['timestamp', 'date', 'time', 'SYS(mmHg)', 'DIA(mmHg)', 'UNKNOW_1', 'UNKNOW_2', 'UNKNOW_3', 'CODE']
         self.data = self.data[order]
 
-        self._eb_dropped = False
-        if (self.data['error'] != 'EB').all():
+        try:
             self.data.set_index('timestamp', inplace=True)
-            self._eb_dropped = True
+        except KeyError:
+            print('Timestamp can not be set as an index:')
+            print(KeyError)
 
         xml_line = open(path, 'r').readlines()[-1]
         xml_root = ET.fromstring(xml_line)
@@ -82,7 +88,7 @@ class SpacelabsReader:
 
     def deidentify(self, subject_id=None):
         """
-        Deidentifies the data by removing the original XML metadata and
+        Deidentify the data by removing the original XML metadata and
         subject id.
 
         Parameters
@@ -108,7 +114,7 @@ class SpacelabsReader:
 
     def write(self, path):
         """
-        Writes the signals and metadata to the
+        Write the signals and metadata to the
         writing path in the same format as it was read.
 
         Parameters
@@ -118,36 +124,36 @@ class SpacelabsReader:
             Use the file extension 'abp' to keep the SpaceLabs standard.
         """
 
-        with open(path, 'w') as f:
-            f.write(f"\n{self.subject}")
-            f.write(8 * '\n')
-            f.write("0")
-            f.write(8 * '\n')
-            f.write(self.data.date[0].strftime("%d.%m.%Y"))
-            f.write(7 * '\n')
-            f.write("Unknown Line")
-            f.write(26 * '\n')
-            f.write(self.valid_measurements + "\n")
[…]
