### 2021-08-13 4b4e67d
made tests compatible to the new version
