+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "devicely"
-version = "1.0.0"
+version = "1.1.0"
 description = "Read multiple signals from different sources."
 authors = [
     "Ariane Morassi Sasso <ariane.morassi-sasso@hpi.de>",
