+++ README.md
@@ -1,4 +1,5 @@
 [![PyPI version](https://badge.fury.io/py/devicely.svg)](https://badge.fury.io/py/devicely)
+[![PyPI pyversions](https://img.shields.io/pypi/pyversions/devicely.svg)](https://pypi.python.org/pypi/devicely/)
 [![Actions Status: test](https://github.com/hpi-dhc/devicely/workflows/test/badge.svg)](https://github.com/hpi-dhc/devicely/actions/workflows/test.yml)
 ![Coverage Badge](https://img.shields.io/endpoint?url=https://gist.githubusercontent.com/jostmorgenstern/270a0114dfad9251945a146dd6d29fa6/raw/devicely_coverage_main.json)
 [![DOI](https://zenodo.org/badge/279395106.svg)](https://zenodo.org/badge/latestdoi/279395106)
