### 2021-08-18 #37 @willingc
I believe that my key recommendations are addressed. Nice improvements 😄
