### 2021-08-18 #37 @arianesasso
@willingc @xmnlab Thank you! We worked hard to improve the documentation, we also added a binder (still improving on that though :)). We are also generating a new version 1.1.0 (probably today). I think the only key suggestion missing is the conda forge feedstock.
