### 2021-08-19 #37 @willingc
@arianesasso Congrats. I've opened a PR for the conda-forge feedstock (https://github.com/conda-forge/staged-recipes/pull/15943). If tests pass, you should be all set there. FYI on the process: I used the grayskull package to generate a `meta.yaml` file which was then submitted as a PR to the conda-forge.
