### 2021-08-19 #37 @arianesasso
Thank you so much @willingc <3, that is really great!!! And @xmnlab which are the next steps now for pyOpenSci and JOSS? :)
