### 2021-08-19 #37 @lwasser
wow this is great - I will let @xmnlab reply separately, but i didn't know about the grayskull package for conda forge @willingc ! that is great to know about and I will check it out. We could then add it to our users / maintainers guide. It's easy to mess up the syntax for that yaml file. I will open an issue in our guide book about this ! @arianesasso congratulations on the reviews being complete!
