### 2021-08-19 #37 @xmnlab
@arianesasso this is great, congrats!

@willingc and @agricolab thank you so much for the review and all the recommendations here.

@arianesasso I will send you the information about the JOSS process soon.

Let me know when you cut a new realease for devicely and I will update here with the new version number.

Thank you so much.
