### 2021-08-20 #37 @lwasser
@xmnlab it looks to me like you nailed it! i removed some of the older labels just so it's clear what stage this package is in. this is minor :) but i think we can assume reviewers were found, etc at this point! I'm so so glad the editors guide is clear. i' spent a lot of time on it and know it probably needs fresh eyes to read it but also know i need to work on the maintainer guide next :) so thank you for that feedback!!

I think now we just let JOSS do it's thing. @arianesasso please be sure to reference this issue when submitting to JOSS so we have a clear record of it. Also please make sure there is a paper.md file following all of JOSS' criteria BEFORE you submit there. I do see that my template above doesn't mention the paper.md file does it? I also see that we may want to talk with JOSS about whether they read the paper or we do or both? regardless @xmnlab you did a fantastic job pushing this through the review process!! Thank you. And thank you again to our reviewers @willingc @agricolab . ALL- we are open to any and all feedback as we improve our review processing and documentation.
