### 2021-08-24 f0dc9f4
Updated the documentation and copyright
