+++ devicely/everion.py
@@ -15,7 +15,7 @@ import pandas as pd
 class EverionReader:
     """
     Read, timeshift and write data generated by Biovotion Everion.
-    You can find example data `here <https://github.com/jostmorgenstern/devicely-documentation-sample-data/tree/main/Everion/>`_.
+    You can find example data `here <https://github.com/hpi-dhc/devicely-example/tree/main/Everion/>`_.
     Each measurement has a tag which is an information about the type of
     measurement. By default, only some of the tags are read for signals, sensors
     and features. You can find these tags via self.default_signal_tags,
