+++ docs/_sources/examples.ipynb.txt
@@ -13,7 +13,7 @@
    "source": [
     "# Use devicely\n",
     "\n",
-    "If you are running this notebook localy don't forget to install the requirements in `requirements.txt`"
+    "To install devicely locally just run `pip install devicely`"
    ]
   },
   {
@@ -25,7 +25,6 @@
     "import os\n",
     "\n",
     "import devicely\n",
-    "import pandas as pd\n",
     "\n",
     "write_dir = 'New'\n",
     "\n",
@@ -279,12 +278,12 @@
     {
      "data": {
       "text/plain": [
-       "{'ACC': Timestamp('2020-04-04 15:47:01.149220446'),\n",
-       " 'BVP': Timestamp('2020-04-04 15:47:01.149220446'),\n",
-       " 'EDA': Timestamp('2020-04-04 15:47:01.149220446'),\n",
-       " 'HR': Timestamp('2020-04-04 15:47:11.149220446'),\n",
-       " 'TEMP': Timestamp('2020-04-04 15:47:01.149220446'),\n",
-       " 'IBI': Timestamp('2020-04-04 15:47:01.149220446')}"
+       "{'ACC': Timestamp('2020-08-07 10:28:47.495956799'),\n",
+       " 'BVP': Timestamp('2020-08-07 10:28:47.495956799'),\n",
+       " 'EDA': Timestamp('2020-08-07 10:28:47.495956799'),\n",
+       " 'HR': Timestamp('2020-08-07 10:28:57.495956799'),\n",
+       " 'TEMP': Timestamp('2020-08-07 10:28:47.495956799'),\n",
+       " 'IBI': Timestamp('2020-08-07 10:28:47.495956799')}"
       ]
      },
      "execution_count": 7,
@@ -612,9 +611,9 @@
        "  </thead>\n",
        "  <tbody>\n",
        "    <tr>\n",
-       "      <th>2018-06-23 01:53:00</th>\n",
-       "      <td>2018-06-23</td>\n",
-       "      <td>01:53:00</td>\n",
+       "      <th>2018-04-12 06:11:00</th>\n",
+       "      <td>2018-04-12</td>\n",
+       "      <td>06:11:00</td>\n",
        "      <td>107</td>\n",
        "      <td>76</td>\n",
        "      <td>78.0</td>\n",
@@ -623,9 +622,9 @@
        "      <td>NaN</td>\n",
        "    </tr>\n",
        "    <tr>\n",
-       "      <th>2018-06-23 01:54:00</th>\n",
-       "      <td>2018-06-23</td>\n",
-       "      <td>01:54:00</td>\n",
+       "      <th>2018-04-12 06:12:00</th>\n",
+       "      <td>2018-04-12</td>\n",
+       "      <td>06:12:00</td>\n",
        "      <td>96</td>\n",
        "      <td>62</td>\n",
        "      <td>63.0</td>\n",
@@ -634,9 +633,9 @@
        "      <td>NaN</td>\n",
        "    </tr>\n",
        "    <tr>\n",
-       "      <th>2018-06-23 01:57:00</th>\n",
-       "      <td>2018-06-23</td>\n",
-       "      <td>01:57:00</td>\n",
+       "      <th>2018-04-12 06:15:00</th>\n",
+       "      <td>2018-04-12</td>\n",
+       "      <td>06:15:00</td>\n",
        "      <td>100</td>\n",
        "      <td>68</td>\n",
        "      <td>64.0</td>\n",
@@ -645,9 +644,9 @@
        "      <td>NaN</td>\n",
        "    </tr>\n",
        "    <tr>\n",
-       "      <th>2018-06-23 01:58:00</th>\n",
[…]
