### 2021-08-24 hpi-dhc/devicely#19 @arianesasso
I believe the tests are now passing and the dependencies are being managed using poetry.
