### 2021-08-24 10c0f56
Update README.md
Updated quick start guide link.
