### 2021-08-24 e99ec8c
Update README.md
Fixed image path for pypi.
