+++ inst/paper.bib
@@ -108,3 +108,11 @@ title = {{FLIRT: Feature generation tooLkIt for weaRable daTa}},
 url = {https://github.com/im-ethz/flirt},
 year = {2020}
 }
+
+@misc{Bent2020,
+author = {Bent, Brinnae},
+booktitle = {DBDP (Digital Biomarker Discovery Pipeline)},
+title = {{wearablecompute is an open source Python package containing over 50 data and domain-driven features}},
+url = {https://github.com/brinnaebent/wearablecompute},
+year = {2020}
+}
