+++ inst/paper.md
@@ -58,15 +58,15 @@ customizable format and is mainly used to extract features for gait analysis.
 Therefore, packages such as SleepPy, HRV and GaitPy could also be used as a
 second step to extract features from the output generated by **devicely**.
 
-FLIRT `[@Maritsch2020:2020]` is the only package that provides ways for
-reading data from specific wearables such as Empatica E4. It also
+FLIRT `[@Maritsch2020:2020]` and wearablecompute `[@Bent2020:2020]` are packages that provides ways for
+reading data from specific wearables such as Empatica E4. They also
 includes functionalities to extract features from electrodermal activity (EDA),
-acceleration and HRV. The main difference from FLIRT to **devicely** is the
+acceleration and HRV. The main difference from FLIRT and wearablecompute to **devicely** is the
 focus on feature extraction versus privacy and data sharing.
-FLIRT reads the data for extracting features, while **devicely** aims to provide users with
-a way to read the data, de-identify them as necessary and write them back. In this
-way, researchers can protect even more data privacy and use the data easily for
-further analysis and sharing.
+FLIRT and wearablecompute read the data for extracting features, while **devicely** aims to provide users with
+a way to read the data, de-identify them as necessary and write them back. In
+this way, researchers can protect even more data privacy and use the data easily
+for further analysis and sharing.
 
 
 # Statement of need
@@ -120,6 +120,14 @@ package and serves as a guide for future implementation.
 
 The software can be obtained through pip and on GitHub at:
 https://github.com/hpi-dhc/devicely under the MIT License.
+
+# Mention
+
+This package was used in the following paper:
+
+Morassi Sasso A. et al. (2020) HYPE: Predicting Blood Pressure from Photoplethysmograms in a Hypertensive Population. In: Michalowski M., Moskovitch R. (eds) Artificial Intelligence in Medicine. AIME 2020. Lecture Notes in Computer Science, vol 12299. Springer, Cham. https://doi.org/10.1007/978-3-030-59137-3_29
+
+GitHub: https://github.com/arianesasso/aime-2020
 
 # Acknowledgements
 
