+++ inst/paper.bib
@@ -104,7 +104,7 @@ year = {2020}
 @misc{Maritsch2020,
 author = {Maritsch, Martin and F{\"{o}}ll, Simon and Fleisch, Elgar and Kowatsch, Tobias and Wortmann, Felix and Spinola, Federica},
 booktitle = {Chair of Information Management at ETH Zurich and the Institute of Technology Management at University of St. Gallen},
-title = {{FLIRT: Feature generation tooLkIt for weaRable daTa}},
+title = {FLIRT: Feature generation tooLkIt for weaRable daTa},
 url = {https://github.com/im-ethz/flirt},
 year = {2020}
 }
@@ -112,7 +112,7 @@ year = {2020}
 @misc{Bent2020,
 author = {Bent, Brinnae},
 booktitle = {DBDP (Digital Biomarker Discovery Pipeline)},
-title = {{wearablecompute is an open source Python package containing over 50 data and domain-driven features}},
+title = {wearablecompute is an open source Python package containing over 50 data and domain-driven features},
 url = {https://github.com/brinnaebent/wearablecompute},
 year = {2020}
 }
