+++ inst/paper.md
@@ -58,7 +58,7 @@ customizable format and is mainly used to extract features for gait analysis.
 Therefore, packages such as SleepPy, HRV and GaitPy could also be used as a
 second step to extract features from the output generated by **devicely**.
 
-FLIRT `[@Maritsch2020:2020]` and wearablecompute `[@Bent2020:2020]` are packages that provides ways for
+FLIRT `[@Maritsch2020:2020]` and wearablecompute `[@Bent2020:2020]` are packages that provide ways for
 reading data from specific wearables such as Empatica E4. They also
 includes functionalities to extract features from electrodermal activity (EDA),
 acceleration and HRV. The main difference from FLIRT and wearablecompute to **devicely** is the
