### 2021-08-24 #37 @arianesasso
Dear @xmnlab, I think I addressed all of the points till the JOSS submission :). We also released 1.1.0. Thank you for all your effort! Also, thanks for the awesome revision and help from @agricolab @willingc and @lwasser! For us the process was very smooth and interesting! We found all the feedback really helpful to improve the quality of the package ;).

ps: I would like to know more ablout: Move Package to PyOpenSci Organization

---

> 🎉 devicely has been approved by pyOpenSci! Thank you @arianesasso for submitting devicely and many thanks to @willingc and @agricolab for reviewing this package! 😸
>
> There are a few things left to do to wrap up this submission:
>
> * [ x ]  Activate Zenodo watching the repo if you haven't already done so.
> * [ x ]  Tag and create a release to create a Zenodo version and DOI.
> * [ x ]  Add the badge for pyOpenSci peer-review to the README.md of devicely. The badge should be `[![pyOpenSci](https://tinyurl.com/y22nb8up)](https://github.com/pyOpenSci/software-review/issues/issue-number)`
> * [ x ]  Add devicely to the pyOpenSci website. @arianesasso, please open a pr to update [this file](https://github.com/pyOpenSci/pyopensci.github.io/blob/main/_data/packages.yml): to add your package and name to the list of contributors
> * [ x ]  @arianesasso, @willingc and @agricolab if you have time and are open to being listed on our website, please add yourselves to [this file](https://github.com/pyOpenSci/pyopensci.github.io/blob/main/_data/contributors.yml) via a pr so we can list you on our website as contributors!
>
> It looks like you would like to submit this package to JOSS. Here are the next steps:
>
> * [ x ]  Login to the JOSS website and fill out the JOSS submission form using your Zenodo DOI. **When you fill out the form, be sure to mention and link to the approved pyOpenSci review.** JOSS will tag your package for expedited review if it is already pyOpenSci approved.
> * [ ]  Wait for a JOSS editor to approve the presubmission (which includes a scope check)
> * [ ]  Once the package is approved by JOSS, you will be given instructions by JOSS about updating the citation information in your README file.
> * [ ]  When the JOSS review is complete, add a comment to your review in the pyOpenSci software-review repo that it has been approved by JOSS.
>
> All -- if you have any feedback for us about the review process please feel free to share it here. We are always looking to improve our process and our documentation in the [contributing-guide](https://www.pyopensci.org/contributing-guide). We have also been updating our documentation to improve the process so all feedback is appreciated!
>
> PS: Optional - Move Package to PyOpenSci Organization (BETA)
>
> rOpenSci packages often live in the rOpenSci organization. PyOpenSci is still figuring out whether this model fits for the Python community. If you are interested in this option, let us know and we can provide more instructions for that.
