+++ inst/paper.bib
@@ -1,52 +1,52 @@
 @article{Kamisalic2018,
-abstract = {Wearable devices have recently received considerable interest due to their great promise for a plethora of applications. Increased research efforts are oriented towards a non-invasive monitoring of human health as well as activity parameters. A wide range of wearable sensors are being developed for real-time non-invasive monitoring. This paper provides a comprehensive review of sensors used in wrist-wearable devices, methods used for the visualization of parameters measured as well as methods used for intelligent analysis of data obtained from wrist-wearable devices. In line with this, the main features of commercial wrist-wearable devices are presented. As a result of this review, a taxonomy of sensors, functionalities, and methods used in non-invasive wrist-wearable devices was assembled.},
-author = {Kamiŝali{\'{c}}, Aida and Fister, Iztok and Turkanovi{\'{c}}, Muhamed and Karakati{\^{c}}, Saŝo},
-doi = {10.3390/s18061714},
-issn = {14248220},
-journal = {Sensors (Switzerland)},
-keywords = {Intelligent analysis,Non-invasive,Sensor,Taxonomy,Visualization,Wrist-wearable device},
-number = {6},
-pmid = {29799504},
-title = {{Sensors and functionalities of non-invasive wrist-wearable devices: A review}},
-url = {www.mdpi.com/journal/sensors},
-volume = {18},
-year = {2018}
+  abstract = {Wearable devices have recently received considerable interest due to their great promise for a plethora of applications. Increased research efforts are oriented towards a non-invasive monitoring of human health as well as activity parameters. A wide range of wearable sensors are being developed for real-time non-invasive monitoring. This paper provides a comprehensive review of sensors used in wrist-wearable devices, methods used for the visualization of parameters measured as well as methods used for intelligent analysis of data obtained from wrist-wearable devices. In line with this, the main features of commercial wrist-wearable devices are presented. As a result of this review, a taxonomy of sensors, functionalities, and methods used in non-invasive wrist-wearable devices was assembled.},
+  author = {Kamiŝali{\'{c}}, Aida and Fister, Iztok and Turkanovi{\'{c}}, Muhamed and Karakati{\^{c}}, Saŝo},
+  doi = {10.3390/s18061714},
+  issn = {14248220},
+  journal = {Sensors (Switzerland)},
+  keywords = {Intelligent analysis,Non-invasive,Sensor,Taxonomy,Visualization,Wrist-wearable device},
+  number = {6},
+  pmid = {29799504},
+  title = {{Sensors and functionalities of non-invasive wrist-wearable devices: A review}},
+  url = {www.mdpi.com/journal/sensors},
+  volume = {18},
+  year = {2018}
 }
 
 @misc{IDC2020,
-author = {IDC},
-booktitle = {2020},
-title = {{Shipments of Wearable Devices Leap to 125 Million Units, Up 35.1{\%} in the Third Quarter, According to IDC}},
-url = {https://www.idc.com/getdoc.jsp?containerId=prUS47067820},
-urldate = {2021-03-15},
-year = {2020}
+  author = {IDC},
+  booktitle = {2020},
+  title = {{Shipments of Wearable Devices Leap to 125 Million Units, Up 35.1{\%} in the Third Quarter, According to IDC}},
+  url = {https://www.idc.com/getdoc.jsp?containerId=prUS47067820},
+  urldate = {2021-03-15},
+  year = {2020}
 }
 
 @article{Bayoumy2021,
[…]
