+++ inst/paper.md
@@ -30,11 +30,11 @@ bibliography: paper.bib
 
 Wearable devices can track a multitude of parameters such as heart rate, body
 temperature, blood oxygen saturation, acceleration, blood glucose and much more
-[@Kamisalic2018:2018]. Moreover, they are becoming increasingly popular with a
-steep increase in market presence in 2020 alone [@IDC2020:2020]. Applications
+[@Kamisalic2018]. Moreover, they are becoming increasingly popular with a
+steep increase in market presence in 2020 alone [@IDC2020]. Applications
 for wearable
-devices vary from tracking cardiovascular risks [@Bayoumy2021:2021] to identifying
-COVID-19 onset [@Mishra2020:2020]. Therefore, there is a great need for scientists to
+devices vary from tracking cardiovascular risks [@Bayoumy2021] to identifying
+COVID-19 onset [@Mishra2020]. Therefore, there is a great need for scientists to
 easily go through data acquired from different wearables and to be able to share
 them while protecting patient privacy.
 In order to solve this problem and empower scientists working with biosignals,
@@ -44,7 +44,7 @@ on what they want: the analysis of biosignals guided by privacy principles.
 
 # Related Work
 
-The first example of a package working with wearable data is mhealthtools [@Snyder2020:2020], which is developed in R and focus on
+The first example of a package working with wearable data is mhealthtools [@Snyder2020], which is developed in R and focus on
 extracting features from sensors such as inertial measurement units (IMUs). Its main difference from **devicely**
 is firstly the language (R versus Python) and secondly their complementary
 nature. Mhealthtools offers functionalities for feature extraction and
@@ -52,15 +52,15 @@ nature. Mhealthtools offers functionalities for feature extraction and
 intended to help users a step before by reading and writing data from
 wearables into standardized formats.
 
-There are also packages developed in python, such as SleepPy [@Christakis2019:2019] which uses raw
+There are also packages developed in python, such as SleepPy [@Christakis2019] which uses raw
 accelerometer data for assessing sleep quantity and quality. The HRV
- [@Bartels2020:2020] package uses csv and text files or python iterables such as lists to generate
-features related to heart rate variability (HRV). GaitPy [@Czech2019:2019] accepts input data in a
+ [@Bartels2020] package uses csv and text files or python iterables such as lists to generate
+features related to heart rate variability (HRV). GaitPy [@Czech2019] accepts input data in a
 customizable format and is mainly used to extract features for gait analysis.
 Therefore, packages such as SleepPy, HRV and GaitPy could also be used as a
 second step to extract features from the output generated by **devicely**.
 
-FLIRT [@Maritsch2020:2020] and wearablecompute [@Bent2020:2020] are packages that provide ways for
+FLIRT [@Maritsch2020] and wearablecompute [@Bent2020] are packages that provide ways for
 reading data from specific wearables such as Empatica E4. They also
 includes functionalities to extract features from electrodermal activity (EDA),
 acceleration and HRV. The main difference from FLIRT and wearablecompute to **devicely** is the
