+++ inst/paper.md
@@ -29,7 +29,7 @@ bibliography: paper.bib
 # Summary
 
 Wearable devices can track a multitude of parameters such as heart rate, body
-temperature, blood oxygen saturation, acceleration, blood glucose and much more
+temperature, blood oxygen saturation, acceleration, blood glucose and many more
 [@Kamisalic2018]. Moreover, they are becoming increasingly popular with a
 steep increase in market presence in 2020 alone [@IDC2020]. Applications
 for wearable
@@ -38,7 +38,7 @@ COVID-19 onset [@Mishra2020]. Therefore, there is a great need for scientists to
 easily go through data acquired from different wearables and to be able to share
 them while protecting patient privacy.
 In order to solve this problem and empower scientists working with biosignals,
-we developed the **devicely** package. It represents the data in a science-friendly
+we developed the **devicely** package. It processes device data into a science-friendly
 format and contains tools for data de-identification. It allows scientists to focus
 on what they want: the analysis of biosignals guided by privacy principles.
 
@@ -52,7 +52,7 @@ nature. Mhealthtools offers functionalities for feature extraction and
 intended to help users a step before by reading and writing data from
 wearables into standardized formats.
 
-There are also packages developed in python, such as SleepPy [@Christakis2019] which uses raw
+There are also packages developed in Python, such as SleepPy [@Christakis2019] which uses raw
 accelerometer data for assessing sleep quantity and quality. The HRV
  [@Bartels2020] package uses csv and text files or python iterables such as lists to generate
 features related to heart rate variability (HRV). GaitPy [@Czech2019] accepts input data in a
@@ -62,12 +62,12 @@ second step to extract features from the output generated by **devicely**.
 
 FLIRT [@Maritsch2020] and wearablecompute [@Bent2020] are packages that provide ways for
 reading data from specific wearables such as Empatica E4. They also
-includes functionalities to extract features from electrodermal activity (EDA),
+include functionalities to extract features from electrodermal activity (EDA),
 acceleration and HRV. The main difference from FLIRT and wearablecompute to **devicely** is the
 focus on feature extraction versus privacy and data sharing.
 FLIRT and wearablecompute read the data for extracting features, while **devicely** aims to provide users with
-a way to read the data, de-identify them as necessary and write them back. In
-this way, researchers can protect even more data privacy and use the data easily
+a way to read the data, de-identify them as necessary and write them back in a specified format. In
+this way, researchers can ensure even more data privacy and use the data easily
 for further analysis and sharing.
 
 # Statement of need
@@ -112,7 +112,7 @@ For all wearables, the written data can be read again using the same reader
 class. Figure 1 depicts the class structure of the **devicely**
 package and serves as a guide for future implementation.
 
-![Figure 1. Shows the structure of the devicely package.](devicely_structure.png)
+![Figure 1. Shows the structure of the devicely package and currently supported devices.](devicely_structure.png)
 
 # Availability
 
@@ -129,7 +129,7 @@ GitHub: https://github.com/arianesasso/aime-2020
 
 # Acknowledgements
 
-We acknowledge contributions from Arpita Kappattanavar, Bjarne Pfitzner, Lin
+We acknowledge contributions from Arpita Kappattanavar, Bjarne Pfitzner, Harry Freitas da Cruz, Lin
 Zhou, Pascal Hecker, Philipp Hildebrandt and Sidratul Moontaha during the genesis and testing of this package.
 
 # References
