### 2021-08-25 #37 @xmnlab
@arianesasso, that sounds great! thank you so much.

About `Move Package to PyOpenSci Organization`, I am checking the steps for that I will be back here with this information soon.

thanks!
