+++ tests/test_empatica.py
@@ -76,9 +76,13 @@ class EmpaticaTestCase(unittest.TestCase):
         )
 
         self.expected_IBI_head = pd.DataFrame(
-            {'seconds_since_start': [145.631666, 146.522332, 151.725695, 152.835121, 153.710161],
-             'IBI': [0.62509, 0.890666, 1.062549, 1.109426, 0.875040]}
-        )
+            {'IBI': [0.62509, 0.890666, 1.062549, 1.109426, 0.875040]},
+            index=pd.to_datetime(['2019-03-01 15:15:01.000000145',
+                                  '2019-03-01 15:15:01.000000146',
+                                  '2019-03-01 15:15:01.000000151',
+                                  '2019-03-01 15:15:01.000000152',
+                                  '2019-03-01 15:15:01.000000153'])
+            )
 
         self.expected_tags = pd.Series(pd.to_datetime([1551453311.68, 1551453318.17, 1551453319.49, 1551453328.04], unit='s'), name='tags')
 
