+++ sphinx/conf.py
@@ -17,8 +17,8 @@ import sphinx_rtd_theme
 # -- Project information -----------------------------------------------------
 
 project = 'devicely'
-copyright = '2020, Ariane Sasso, Jost Morgenstern, Felix Musmann, Bert Arnrich'
-author = 'Ariane Sasso, Jost Morgenstern'
+copyright = '2020, Digital Health Center (Hasso Plattner Institute)'
+author = 'Ariane Sasso, Jost Morgenstern, Felix Musmann, Bert Arnrich'
 
 # The full version, including alpha/beta/rc tags
 release = '1.1.0'
