### 2021-09-07 404be09
Solves the compatibility issue with 0.2.5 and empatica ACC files
