+++ inst/paper.md
@@ -49,16 +49,14 @@ extracting features from sensors such as inertial measurement units (IMUs). Its 
 is firstly the language (R versus Python) and secondly their complementary
 nature. Mhealthtools offers functionalities for feature extraction and
 **devicely** is
-intended to help users a step before by reading and writing data from
+intended to help users in a prior step by reading and writing data from
 wearables into standardized formats.
 
 There are also packages developed in Python, such as SleepPy [@Christakis2019] which uses raw
-accelerometer data for assessing sleep quantity and quality. The HRV
- [@Bartels2020] package uses csv and text files or python iterables such as lists to generate
+accelerometer data for assessing sleep quantity and quality. The HRV [@Bartels2020] package uses CSV and text files or Python iterables such as lists to generate
 features related to heart rate variability (HRV). GaitPy [@Czech2019] accepts input data in a
 customizable format and is mainly used to extract features for gait analysis.
-Therefore, packages such as SleepPy, HRV and GaitPy could also be used as a
-second step to extract features from the output generated by **devicely**.
+Therefore, packages such as SleepPy, HRV and GaitPy could also be used as a following step to extract features from the output generated by **devicely**.
 
 FLIRT [@Maritsch2020] and wearablecompute [@Bent2020] are packages that provide ways for
 reading data from specific wearables such as Empatica E4. They also
@@ -72,7 +70,7 @@ for further analysis and sharing.
 
 # Statement of need
 
-Every wearable company has a different data format and reading them is
+Every wearable vendor has a different data format and reading them is
 usually a challenge for scientists. Therefore, in order for researchers to be able to use different sensor data in an easy and
 friendly way we developed the **devicely** package. The package also contains two methods to help with data de-identification, one
 is called timeshift and the other is a write method. The idea behind them is
@@ -83,7 +81,7 @@ and hopefully share more data to increase research reproducibility.
 
 # Design
 
-Different wearables come with different data formats which require different preprocessing steps.
+Different wearables provide incompatible data formats which require specific preprocessing steps.
 However, it should be easy for scientists to add data from a new wearable to an
 existing pipeline and easy for developers to add a new wearable to the
 **devicely** package. To achieve both **devicely** encapsulates data preparation for
