### 2021-09-08 8c118a1
Solves #39
