+++ inst/paper.md
@@ -36,7 +36,7 @@ for wearable
 devices vary from tracking cardiovascular risks [@Bayoumy2021] to identifying
 COVID-19 onset [@Mishra2020]. Therefore, there is a great need for scientists to
 easily go through data acquired from different wearables and to be able to share
-them while protecting patient privacy.
+them while protecting user privacy.
 In order to solve this problem and empower scientists working with biosignals,
 we developed the **devicely** package. It processes the data into a science-friendly
 format and contains tools for data de-identification. It allows scientists to focus
@@ -56,7 +56,7 @@ There are also packages developed in Python, such as SleepPy [@Christakis2019] w
 accelerometer data for assessing sleep quantity and quality. The HRV [@Bartels2020] package uses CSV and text files or Python iterables such as lists to generate
 features related to heart rate variability (HRV). GaitPy [@Czech2019] accepts input data in a
 customizable format and is mainly used to extract features for gait analysis.
-Therefore, packages such as SleepPy, HRV and GaitPy could also be used as a following step to extract features from the output generated by **devicely**.
+Therefore, packages such as SleepPy, HRV and GaitPy could also be used in a following step to extract features from the output generated by **devicely**.
 
 FLIRT [@Maritsch2020] and wearablecompute [@Bent2020] are packages that provide ways for
 reading data from specific wearables such as Empatica E4. They also
@@ -76,7 +76,7 @@ friendly way we developed the **devicely** package. The package also contains tw
 is called timeshift and the other is a write method. The idea behind them is
 that researchers can timeshift all their time series data to a different time from the one the
 actual experiments occurred and then write this new de-identified dataset back to
-the original or a similar data format. This will empower scientists to keep patient privacy
+the original or a similar data format. This will empower scientists to keep user privacy
 and hopefully share more data to increase research reproducibility.
 
 # Design
@@ -91,7 +91,7 @@ After reading, the data is accessible through the reader in common formats such 
 De-identification is achieved by timeshifting the data, either by providing a shifting interval or randomly.
 For writing back de-identified data **devicely** focuses on keeping a format that can be read again using the same reader class.
 In almost all cases, this is the same format as the one the wearable originaly provides.
-This enables sharing data with the community while maintaining patient privacy.
+This enables sharing data with the community while maintaining user privacy.
 
 # Functionalities
 
