### 2021-09-08 a04dfbc
Suggestion for #41
