### 2021-09-09 beb1232
Solves #44
