### 2021-09-09 7a5b411
Solves #49
