+++ devicely/empatica.py
@@ -164,10 +164,11 @@ class EmpaticaReader:
                 with open(path, 'r') as file:
                     start_time = pd.Timestamp(float(file.readline().split(',')[0]), unit='s')
                     self.start_times['IBI'] = start_time
-                    dataframe = pd.read_csv(file, names=['time', 'IBI'], header=None)
-                    
-
-                    return dataframe
+                    df = pd.read_csv(file, names=['time', 'IBI'], header=None)
+                    df['time'] = pd.to_timedelta(df['time'], unit='s')
+                    df['time'] = start_time + df['time']
+#                    breakpoint()
+                    return df.set_index('time') 
             else:
                 print(f"Not reading signal because the file {path} is empty.")
         except OSError:
@@ -177,8 +178,11 @@ class EmpaticaReader:
 
     def _write_ibi(self, path):
         with open(path, 'w') as file:
-            file.write(f"{float(self.start_times['IBI'].value / 1e9)}, IBI\n")
-            file.write(self.IBI.to_csv(index=None, header=None))
+            file.write(f"{self.start_times['IBI'].value // 1e9}, IBI\n")
+            write_df = self.IBI.copy()
+            write_df.index = (write_df.index - self.start_times['IBI']).values.astype(int) / 1e9
+            write_df.to_csv(file, header=None)
+
 
     def _read_tags(self, path):
         try:
