+++ tests/test_empatica.py
@@ -77,11 +77,9 @@ class EmpaticaTestCase(unittest.TestCase):
 
         self.expected_IBI_head = pd.DataFrame(
             {'IBI': [0.62509, 0.890666, 1.062549, 1.109426, 0.875040]},
-            index=pd.to_datetime(['2019-03-01 15:15:01.000000145',
-                                  '2019-03-01 15:15:01.000000146',
-                                  '2019-03-01 15:15:01.000000151',
-                                  '2019-03-01 15:15:01.000000152',
-                                  '2019-03-01 15:15:01.000000153'])
+            index=pd.DatetimeIndex(['2019-03-01 15:17:26.631666', '2019-03-01 15:17:27.522332',
+                                    '2019-03-01 15:17:32.725695', '2019-03-01 15:17:33.835121',
+                                    '2019-03-01 15:17:34.710161'], name='time') 
             )
 
         self.expected_tags = pd.Series(pd.to_datetime([1551453311.68, 1551453318.17, 1551453319.49, 1551453328.04], unit='s'), name='tags')
@@ -104,6 +102,7 @@ class EmpaticaTestCase(unittest.TestCase):
         pd.testing.assert_series_equal(reader.EDA.head(), expected_EDA_head, check_dtype=False, check_freq=False)
         pd.testing.assert_series_equal(reader.HR.head(), expected_HR_head, check_dtype=False, check_freq=False)
         pd.testing.assert_series_equal(reader.TEMP.head(), expected_TEMP_head, check_dtype=False, check_freq=False)
+#        breakpoint()
         pd.testing.assert_frame_equal(reader.IBI.head(), expected_IBI_head, check_dtype=False, check_freq=False)
         pd.testing.assert_series_equal(reader.tags, expected_tags, check_freq=False)
 
@@ -117,7 +116,6 @@ class EmpaticaTestCase(unittest.TestCase):
 
     def test_read(self):
         # tests the basic reading capability by comparing the read data to the expected values
-
         self._test_compare_real_and_expected(self.reader, self.expected_start_times, self.expected_sample_freqs,
                                                           self.expected_ACC_head, self.expected_BVP_head,
                                                           self.expected_EDA_head, self.expected_HR_head,
@@ -128,6 +126,7 @@ class EmpaticaTestCase(unittest.TestCase):
         # test the writing capability by writing, reading with a new reader and comparing the old to the new data
 
         self.reader.write(self.WRITE_PATH)
+       # breakpoint()
         new_reader = devicely.EmpaticaReader(self.WRITE_PATH)
         self._test_compare_real_and_expected(new_reader, self.expected_start_times, self.expected_sample_freqs,
                                                          self.expected_ACC_head, self.expected_BVP_head,
