### 2021-09-10 fda3e9c
make faros join optional, docs and binder repo adjusted accordingly
