+++ docs/_sources/examples.ipynb.txt
@@ -23,7 +23,7 @@
   },
   {
    "cell_type": "code",
-   "execution_count": 1,
+   "execution_count": 85,
    "metadata": {},
    "outputs": [],
    "source": [
@@ -57,7 +57,7 @@
   },
   {
    "cell_type": "code",
-   "execution_count": 2,
+   "execution_count": 86,
    "metadata": {},
    "outputs": [],
    "source": [
@@ -73,7 +73,7 @@
   },
   {
    "cell_type": "code",
-   "execution_count": 3,
+   "execution_count": 87,
    "metadata": {},
    "outputs": [
     {
@@ -87,7 +87,7 @@
        " 'IBI': Timestamp('2020-10-16 17:04:29')}"
       ]
      },
-     "execution_count": 3,
+     "execution_count": 87,
      "metadata": {},
      "output_type": "execute_result"
     }
@@ -98,7 +98,7 @@
   },
   {
    "cell_type": "code",
-   "execution_count": 4,
+   "execution_count": 88,
    "metadata": {},
    "outputs": [
     {
@@ -107,7 +107,7 @@
        "{'ACC': 32.0, 'BVP': 64.0, 'EDA': 4.0, 'HR': 1.0, 'TEMP': 4.0}"
       ]
      },
-     "execution_count": 4,
+     "execution_count": 88,
      "metadata": {},
      "output_type": "execute_result"
     }
@@ -125,7 +125,7 @@
   },
   {
    "cell_type": "code",
-   "execution_count": 5,
+   "execution_count": 89,
    "metadata": {},
    "outputs": [
     {
@@ -139,7 +139,7 @@
        "Name: HR, dtype: float64"
       ]
      },
-     "execution_count": 5,
+     "execution_count": 89,
      "metadata": {},
      "output_type": "execute_result"
     }
@@ -157,7 +157,7 @@
   },
   {
    "cell_type": "code",
-   "execution_count": 6,
+   "execution_count": 90,
    "metadata": {},
    "outputs": [
     {
@@ -254,7 +254,7 @@
        "2020-10-16 17:04:29.062500   -2.0   11.0   60.0 -0.0  NaN NaN    NaN"
       ]
      },
-     "execution_count": 6,
+     "execution_count": 90,
      "metadata": {},
      "output_type": "execute_result"
     }
@@ -277,21 +277,21 @@
   },
   {
    "cell_type": "code",
-   "execution_count": 7,
+   "execution_count": 91,
    "metadata": {},
    "outputs": [
     {
      "data": {
       "text/plain": [
-       "{'ACC': Timestamp('2018-11-02 01:32:42.176015200'),\n",
-       " 'BVP': Timestamp('2018-11-02 01:32:42.176015200'),\n",
-       " 'EDA': Timestamp('2018-11-02 01:32:42.176015200'),\n",
-       " 'HR': Timestamp('2018-11-02 01:32:52.176015200'),\n",
-       " 'TEMP': Timestamp('2018-11-02 01:32:42.176015200'),\n",
-       " 'IBI': Timestamp('2018-11-02 01:32:42.176015200')}"
+       "{'ACC': Timestamp('2018-11-14 13:20:16.439132912'),\n",
+       " 'BVP': Timestamp('2018-11-14 13:20:16.439132912'),\n",
+       " 'EDA': Timestamp('2018-11-14 13:20:16.439132912'),\n",
+       " 'HR': Timestamp('2018-11-14 13:20:26.439132912'),\n",
+       " 'TEMP': Timestamp('2018-11-14 13:20:16.439132912'),\n",
+       " 'IBI': Timestamp('2018-11-14 13:20:16.439132912')}"
       ]
      },
-     "execution_count": 7,
+     "execution_count": 91,
[…]
