+++ .github/workflows/docs.yml
@@ -1,0 +1,24 @@
+name: "Sphinx Build"
+on:
+  push:
+    branches:
+    - test-ci-pypi
+jobs:
+  docs:
+    runs-on: ubuntu-latest
+    steps:
+    - uses: actions/checkout@v2
+    
+    - name: Set up Python
+      uses: actions/setup-python@v2
+      with:
+        python-version: 3.9
+
+    - name: Install Poetry
+      uses: snok/install-poetry@v1
+
+    - name: Install dependencies
+      run: |
+        poetry install
+        poetry build
+
