+++ .github/workflows/pypi.yml
@@ -1,8 +1,8 @@
-name: "Sphinx Build"
+name: "Publish to PyPi"
 on:
   push:
     branches:
-    - test-ci-pypi
+    - CI-testpypi
 jobs:
   docs:
     runs-on: ubuntu-latest
@@ -17,8 +17,14 @@ jobs:
     - name: Install Poetry
       uses: snok/install-poetry@v1
 
-    - name: Install dependencies
+    - name: Build artifacts
       run: |
         poetry install
         poetry build
+
+    - name: Publish dists
+      uses: pypa/gh-action-pypi-publish@main
+      with:
+        password: ${{ secrets.TEST_PYPI_API_TOKEN }}
+        repository_url: https://test.pypi.org/legacy/
 
