+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "devicely"
-version = "1.1.2"
+version = "1.1.3"
 description = "Devicely: A Python package for reading, timeshifting and writing sensor data."
 authors = [
     "Ariane Morassi Sasso <ariane.morassi-sasso@hpi.de>",
