+++ .github/workflows/pypi.yml
@@ -23,7 +23,7 @@ jobs:
         poetry build
 
     - name: Publish dists
-      uses: pypa/gh-action-pypi-publish@main
+      uses: pypa/gh-action-pypi-publish
       with:
         password: ${{ secrets.TEST_PYPI_API_TOKEN }}
         repository_url: https://test.pypi.org/legacy/
