+++ .github/workflows/pypi.yml
@@ -1,4 +1,4 @@
-name: "Publish to PyPi"
+name: "Publish to Test PyPi"
 on:
   push:
     branches:
