+++ .github/workflows/pypi.yml
@@ -23,8 +23,9 @@ jobs:
         poetry build
 
     - name: Publish dists
-      uses: pypa/gh-action-pypi-publish
+      uses: pypa/gh-action-pypi-publish/release/v1
       with:
+        user: __token__
         password: ${{ secrets.TEST_PYPI_API_TOKEN }}
         repository_url: https://test.pypi.org/legacy/
 
