+++ .github/workflows/pypi.yml
@@ -23,7 +23,7 @@ jobs:
         poetry build
 
     - name: Publish dists
-      uses: pypa/gh-action-pypi-publish/release/v1
+      uses: pypa/gh-action-pypi-publish@release/v1
       with:
         user: __token__
         password: ${{ secrets.TEST_PYPI_API_TOKEN }}
