### 2021-09-15 d114bf7
change pypi workflow trigger from push to release
