+++ .github/workflows/pypi.yml
@@ -1,8 +1,7 @@
-name: "Publish to Test PyPi"
+name: "Publish to PyPi"
 on:
-  push:
-    branches:
-    - CI-testpypi
+  release:
+    types: [published]
 jobs:
   docs:
     runs-on: ubuntu-latest
