+++ .github/workflows/docs.yml
@@ -32,6 +32,10 @@ jobs:
 
     - name: Push built docs
       run: |
+        git config --global user.name 'Jost Morgenstern'
+        git config --global user.email 'jostmorgenstern@users.noreply.github.com'
+        git commit -am "automatic docs generation on release"
+        git push
         
 
 
