+++ .github/workflows/pypi.yml
@@ -2,6 +2,8 @@ name: "Publish to PyPi"
 on:
   release:
     types: [published]
+    branches:
+      - main
 jobs:
   docs:
     runs-on: ubuntu-latest
