### 2021-09-25 2f51a6a
create global version number truth in pyproject.toml
