+++ sphinx/conf.py
@@ -12,7 +12,7 @@
 #
 import sphinx_rtd_theme
 
-
+import devicely
 
 # -- Project information -----------------------------------------------------
 
@@ -21,7 +21,7 @@ copyright = '2021, Digital Health Center (Hasso Plattner Institute)'
 author = 'Ariane Sasso, Jost Morgenstern, Felix Musmann, Bert Arnrich'
 
 # The full version, including alpha/beta/rc tags
-release = '1.1.0'
+release = devicely.__version__
 
 
 # -- General configuration ---------------------------------------------------
