+++ pyproject.toml
@@ -25,6 +25,7 @@ python = "^3.7.1"
 pandas = "^1.3.0"
 pyEDFlib = "^0.1.22"
 numpy = "^1.21.1"
+importlib-metadata = {version = "^1.0", python = "<3.8"}
 
 [tool.poetry.dev-dependencies]
 pytest = "^6.2.4"
