+++ tests/test_muse.py
@@ -1,0 +1,105 @@
+"""
+Tests for the muse module
+"""
+import os
+import sys
+import unittest
+
+import numpy as np
+import pandas as pd
+
+import devicely
+
+
+class MuseTestCase(unittest.TestCase):
+    READ_PATH = 'tests/Muse_test_data/read_test_data.csv'
+    WRITE_PATH = 'tests/Muse_test_data/write_test_data.csv'
+
+    expected_data_head = pd.DataFrame(
+        {'AUX_RIGHT': [844.1392, np.nan, 861.06226, 1009.34064, 1006.52014],
+         'Accelerometer_X': [-0.23260498046875, np.nan, -0.23260498046875, -0.23260498046875, -0.23260498046875],
+         'Accelerometer_Y': [-0.04901123046875, np.nan, -0.04901123046875, -0.04901123046875, -0.04901123046875],
+         'Accelerometer_Z': [0.9664306640625, np.nan, 0.9664306640625, 0.9664306640625, 0.9664306640625],
+         'Alpha_AF7': [0.70684004, np.nan, 0.7046354, 0.7046354, 0.7046354],
+         'Alpha_AF8': [0.3738289, np.nan, 0.37960157, 0.37960157, 0.37960157],
+         'Alpha_TP10': [0.5247787, np.nan, 0.5247787, 0.5247787, 0.5247787],
+         'Alpha_TP9': [0.6818201, np.nan, 0.6818201, 0.6818201, 0.6818201],
+         'Battery': [100.0, np.nan, 100.0, 100.0, 100.0],
+         'Beta_AF7': [0.9047759, np.nan, 0.90967786, 0.90967786, 0.90967786],
+         'Beta_AF8': [-0.074241824, np.nan, -0.07182129, -0.07182129, -0.07182129],
+         'Beta_TP10': [0.24255054, np.nan, 0.24255054, 0.24255054, 0.24255054],
+         'Beta_TP9': [0.40113458, np.nan, 0.40113458, 0.40113458, 0.40113458],
+         'Delta_AF7': [0.65365446, np.nan, 0.6321334, 0.6321334, 0.6321334],
+         'Delta_AF8': [0.87098247, np.nan, 0.8560632, 0.8560632, 0.8560632],
+         'Delta_TP10': [0.7412762, np.nan, 0.7412762, 0.7412762, 0.7412762],
+         'Delta_TP9': [0.7765235, np.nan, 0.7765235, 0.7765235, 0.7765235],
+         'Elements': [np.nan, '/muse/elements/blink', np.nan, np.nan, np.nan],
+         'Gamma_AF7': [0.7472885, np.nan, 0.73841846, 0.73841846, 0.73841846],
+         'Gamma_AF8': [-0.29481375, np.nan, -0.29959682, -0.29959682, -0.29959682],
+         'Gamma_TP10': [-0.3348679, np.nan, -0.3348679, -0.3348679, -0.3348679],
+         'Gamma_TP9': [-0.0852012, np.nan, -0.0852012, -0.0852012, -0.0852012],
+         'Gyro_X': [2.998199462890625, np.nan, 2.998199462890625, 2.998199462890625, 2.998199462890625],
+         'Gyro_Y': [2.130889892578125, np.nan, 2.130889892578125, 2.130889892578125, 2.130889892578125],
+         'Gyro_Z': [1.966400146484375, np.nan, 1.966400146484375, 1.966400146484375, 1.966400146484375],
+         'HSI_AF7': [1.0, np.nan, 1.0, 1.0, 1.0],
[…]
