### 2021-09-28 0ca1d1e
add Muse to sphinx docs
