### 2021-09-29 982e692
add muse description to README and docs
