### 2021-10-11 3ba16e1
Corrections suggested by JOSS editor
