### 2021-10-20 7d94571
change docs release workflow
