+++ .github/workflows/docs.yml
@@ -34,9 +34,7 @@ jobs:
         touch docs/.nojekyll
 
     - name: Push built docs
-      run: |
-        git add .
-        git config --global user.name 'Jost Morgenstern'
-        git config --global user.email 'jostmorgenstern@users.noreply.github.com'
-        git commit -m "automatic docs generation on release"
-        git push
+      uses: EndBug/add-and-commit@v7
+      with:
+        message: Automatic documentation generation on release
+
