+++ .github/workflows/pypi.yml
@@ -5,7 +5,7 @@ on:
     branches:
       - main
 jobs:
-  docs:
+  publish-dists:
     runs-on: ubuntu-latest
     steps:
     - uses: actions/checkout@v2
