+++ pyproject.toml
@@ -38,3 +38,4 @@ sphinx-rtd-theme = "^0.5.2"
 [build-system]
 requires = ["poetry-core>=1.0.0"]
 build-backend = "poetry.core.masonry.api"
+
