+++ inst/paper.bib
@@ -101,12 +101,12 @@
   journal = {Journal of Open Source Software}
 }
 
-@misc{Maritsch2020,
-  author = {Maritsch, Martin and F{\"{o}}ll, Simon and Fleisch, Elgar and Kowatsch, Tobias and Wortmann, Felix and Spinola, Federica},
-  publisher = {Chair of Information Management at ETH Zurich and the Institute of Technology Management at University of St. Gallen},
-  title = {FLIRT: Feature generation tooLkIt for weaRable daTa},
-  url = {https://github.com/im-ethz/flirt},
-  year = {2020}
+@article{Maritsch2020,
+  doi = {10.1016/j.cmpb.2021.106461},
+  author = {Föll, Simon and Maritsch, Martin and Spinola, Federica and Mishra, Varun and Barata, Filipe and Kowatsch, Tobias and Fleisch, Elgar and Wortmann, Felix},
+  title = {FLIRT: A Feature Generation Toolkit for Wearable Data},
+  journal = {Computer Methods and Programs in Biomedicine},
+  year = {2021}
 }
 
 @misc{Bent2020,
