+++ inst/paper.bib
@@ -101,10 +101,10 @@
   journal = {Journal of Open Source Software}
 }
 
-@article{Maritsch2020,
+@article{Foell2020,
   doi = {10.1016/j.cmpb.2021.106461},
   author = {Föll, Simon and Maritsch, Martin and Spinola, Federica and Mishra, Varun and Barata, Filipe and Kowatsch, Tobias and Fleisch, Elgar and Wortmann, Felix},
-  title = {FLIRT: A Feature Generation Toolkit for Wearable Data},
+  title = {{{FLIRT}}: A {{Feature Generation Toolkit}} for {{Wearable Data}}},
   journal = {Computer Methods and Programs in Biomedicine},
   year = {2021}
 }
