+++ .github/workflows/docs.yml
@@ -1,8 +1,7 @@
 name: "Sphinx Build"
 on:
-  release:
-    types: [published]
-    branches:
+  push:
+    branch:
       - main
 jobs:
   docs:
@@ -26,6 +25,8 @@ jobs:
     - name: Build documentation
       run: |
         cd sphinx
+        poetry run make clean
+        ls _build/html
         poetry run make html
         ls _build/html
         cd ..
@@ -33,8 +34,17 @@ jobs:
         mv sphinx/_build/html docs
         touch docs/.nojekyll
 
-    - name: Push built docs
-      uses: EndBug/add-and-commit@v7
-      with:
-        message: Automatic documentation generation on release
+        #     - name: Push built docs
+        #      uses: EndBug/add-and-commit@v7
+        #with:
+        #message: Automatic documentation generation on release
+        #
+     - name: Push built docs
+       run: |
+         git config --global user.name 'jostmorgenstern'
+         git config --global user.email 'jostmorgenstern@users.noreply.github.com'
+         git add .
+         git commit -am "Automated docs generation"
+         git push origin main
+
 
