+++ .github/workflows/docs.yml
@@ -34,12 +34,12 @@ jobs:
         mv sphinx/_build/html docs
         touch docs/.nojekyll
 
-     - name: Push built docs
-       run: |
-         git config --global user.name 'jostmorgenstern'
-         git config --global user.email 'jostmorgenstern@users.noreply.github.com'
-         git add .
-         git commit -am "Automated docs generation"
-         git push origin main
+    - name: Push built docs
+      run: |
+        git config --global user.name 'jostmorgenstern'
+        git config --global user.email 'jostmorgenstern@users.noreply.github.com'
+        git add .
+        git commit -am "Automated docs generation"
+        git push origin main
 
 
