+++ .github/workflows/docs.yml
@@ -26,14 +26,12 @@ jobs:
       run: |
         cd sphinx
         poetry run make clean
-        ls
         poetry run make html
-        ls
-        ls _build/html
         cd ..
         rm -rf docs
         mv sphinx/_build/html docs
         touch docs/.nojekyll
+        ls -la docs
 
     - name: Push built docs
       run: |
@@ -41,6 +39,7 @@ jobs:
         git config --global user.email 'jostmorgenstern@users.noreply.github.com'
         git add .
         git commit -am "Automated docs generation"
+        git status
         git push origin main
 
 
