### 2021-12-09 10c87c6
Update README.md
Added citation
