### 2022-09-01 #37 @lwasser
hi there @arianesasso !! i am checkin in on this review. Did you end up moving forward with submitting to JOSS to get your cross-ref validated DOI? If you decided against it I am going to close this issue and remove the JOSS tag. If you did - can you kindly send us a link to the JOSS issue? Many thanks.

In the future we do plan to discuss the option of moving repos over to the pyopensci organization in the same way that ropensci does so that input is noted. So far most of our submissions have been happy with remaining in their own organizations but that doesn't mean that all will want to be treated that way.
