### 2022-09-05 #37 @arianesasso
Hi @lwasser, we did publish it on JOSS :). Here is the link: https://joss.theoj.org/papers/10.21105/joss.03679

Thanks for letting us know! For now, we are also ok with having it in our organization.

Cheers!
